% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.EffortTimeSeries.R
\name{compareData.EffortTimeSeries}
\alias{compareData.EffortTimeSeries}
\title{Compare effort time series data by fleet among several model scenarios}
\usage{
compareData.EffortTimeSeries(objs = NULL, scales = "free_y",
  verbose = FALSE)
}
\arguments{
\item{scales}{- ggplot2 scales option for facet_grid}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare effort time series data by fleet among 
several model scenarios.
}
\details{
Uses \code{rTCSAM02::getMDFR.Data.EffortData()}.
}
