% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.FleetTimeSeries.R
\name{compareData.FleetTimeSeries}
\alias{compareData.FleetTimeSeries}
\title{Compare biomass time series data by fleet among several model scenarios}
\usage{
compareData.FleetTimeSeries(objs = NULL, fleet.type = c("survey",
  "fishery"), catch.type = c("index", "retained", "discard", "total"),
  data.type = c("abundance", "biomass"), ci = 0.8, numRecent = 15,
  facets = "fleet~x", scales = "free_y", verbose = FALSE)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained', 'discard' or 'total')}

\item{data.type}{- data type ('abundance' or 'biomass')}

\item{ci}{- confidence interval for plots}

\item{numRecent}{- number of years for 'recent' plot}

\item{facets}{- grid faceting formula}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare biomass time series data by fleet among 
several model scenarios.
}
\details{
Uses \code{rTCSAM02::getMDFR.Data.FleetTimeSeries()}.
}
