% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.FleetTimeSeriesABs.R
\name{compareData.FleetTimeSeriesABs}
\alias{compareData.FleetTimeSeriesABs}
\title{Compare time series of abundance or biommass data by fleet among several model scenarios}
\usage{
compareData.FleetTimeSeriesABs(
  objs = NULL,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  data.type = c("abundance", "biomass"),
  fleets = "all",
  sexs = "all",
  maturity_states = "all",
  shell_conditions = "all",
  ci = 0.8,
  numRecent = 15,
  ylims = NULL,
  facets = "fleet~x",
  scales = "free_y",
  verbose = FALSE
)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained', 'discard' or 'total')}

\item{data.type}{- data type ('abundance' or 'biomass')}

\item{fleets}{- vector of names of fleets to plot (or 'all')}

\item{sexs}{- vector of sexes to plot (or 'all')}

\item{maturity_states}{- vector of maturity states to plot (or 'all')}

\item{shell_conditions}{- vector of shell conditions to plot (or 'all')}

\item{ci}{- confidence interval for plots}

\item{numRecent}{- number of years for 'recent' plot}

\item{ylims}{- limits for y axis (default is NULL)}

\item{facets}{- grid faceting formula}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare abundance or biomass time series data by fleet among
several model scenarios.
}
\details{
Uses \code{rTCSAM02::getMDFR.Data.FleetTimeSeries()}.
}
