% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.GrowthData.R
\name{compareData.GrowthData}
\alias{compareData.GrowthData}
\title{Compare growth data among several model scenarios}
\usage{
compareData.GrowthData(objs, dodge = 0.2, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot object
}
\description{
Function to compare growth data among several model scenarios.
}
\details{
None.
}
