% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.EffectiveNs.R
\name{compareFits.EffectiveNs}
\alias{compareFits.EffectiveNs}
\title{Function to compare effective Ns from size compositions by year among several models}
\usage{
compareFits.EffectiveNs(
  objs = NULL,
  mdfr = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  category = c("index", "captured", "discarded", "retained", "discard mortality"),
  years = "all",
  dodge = 0.2,
  facet_wrap = NULL,
  facet_grid = "x+m+s~.",
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{mdfr}{\itemize{
\item dataframe
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all" or NULL to include all)
}}

\item{fleet.type}{\itemize{
\item 'survey','fishery'
}}

\item{category}{\itemize{
\item 'total','discard','retained','discard mortality', or 'index'
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{facet_wrap}{\itemize{
\item ggplot2 formula to produce figure with wrapped facets
}}

\item{facet_grid}{\itemize{
\item ggplot2 formula to produce figure with gridded facets
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
lists ofggplot2 objects, nested by fishery
}
\description{
This function compares effective Ns from size compositions by year
among several models.
}
\details{
None.
}
