% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.MeanSizeComps.R
\name{compareFits.MeanSizeComps}
\alias{compareFits.MeanSizeComps}
\title{Compare fits to mean size comps by fleet among several model runs}
\usage{
compareFits.MeanSizeComps(
  objs = NULL,
  mdfr = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  years = "all",
  ci = 0.8,
  plot1stObs = TRUE,
  facet_grid = "m+s~x",
  scales = "free_y",
  linesize = 1,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}

\item{mdfr}{\itemize{
\item melted dataframe from call to \code{\link[=extractFits.MeanSizeComps]{extractFits.MeanSizeComps()}} (as alternative to objs)
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained',  or 'total')
}}

\item{years}{\itemize{
\item years to plot, as numerical vector (or "all" to plot all years)
}}

\item{ci}{\itemize{
\item confidence interval
}}

\item{plot1stObs}{\itemize{
\item flag (T/F) to plot observations only from first case
}}

\item{facet_grid}{\itemize{
\item faceting formula for \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}
}}

\item{scales}{\itemize{
\item scales parameter for \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}
}}

\item{linesize}{\itemize{
\item line width (in mm)
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list of ggplot2 objects, with captions as names
}
\description{
Function to compare fits to mean size comps by fleet among
several model runs.
}
\details{
Uses \code{\link[=extractFits.MeanSizeComps]{extractFits.MeanSizeComps()}} to extract data.
Also uses \code{\link[reshape2:cast]{reshape2::dcast()}}, \code{\link[wtsUtilities:calcCIs]{wtsUtilities::calcCIs()}}, \code{\link[wtsUtilities:printGGList]{wtsUtilities::printGGList()}}.
}
