% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.SizeComps.R
\name{compareFits.SizeComps}
\alias{compareFits.SizeComps}
\title{Compare fits to size comps by fleet among several model runs}
\usage{
compareFits.SizeComps(
  objs = NULL,
  mdfr = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  years = "all",
  plot1stObs = TRUE,
  nrow = 5,
  ncol = 4,
  useBars = TRUE,
  usePins = FALSE,
  usePinsAndPts = FALSE,
  useLines = TRUE,
  usePoints = TRUE,
  pinSize = 0.2,
  lineSize = 1,
  pointSize = 1,
  alpha = 0.5,
  stripText = ggplot2::element_text(),
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}

\item{mdfr}{\itemize{
\item dataframe from call to \code{\link[=extractFits.SizeComps]{extractFits.SizeComps()}} (as alternative to objs)
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained',  or 'total')
}}

\item{years}{\itemize{
\item years to plot, as numerical vector (or "all" to plot all years)
}}

\item{plot1stObs}{\itemize{
\item flag (T/F) to plot observations only from first case, or character vector cases cases from which to plot observations
}}

\item{nrow}{\itemize{
\item number of rows per page for output plots
}}

\item{ncol}{\itemize{
\item number of columns per page for output plots
}}

\item{useBars}{\itemize{
\item flag to use bars for observations
}}

\item{usePins}{\itemize{
\item flag to use pins for observations
}}

\item{usePinsAndPts}{\itemize{
\item flag to add pts to observations when pins are used
}}

\item{useLines}{\itemize{
\item flag to use lines for predictions
}}

\item{usePoints}{\itemize{
\item flag to use points for predictions
}}

\item{pinSize}{\itemize{
\item width of pin line
}}

\item{lineSize}{\itemize{
\item prediction line size
}}

\item{pointSize}{\itemize{
\item prediction point size
}}

\item{alpha}{\itemize{
\item prediction transparency
}}

\item{stripText}{\itemize{
\item \code{\link[ggplot2:element]{ggplot2::element_text()}} object describing font and margin to use for panel strips
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare fits to size comps by fleet among
several model runs.
}
\details{
Uses \code{\link[=extractFits.SizeComps]{extractFits.SizeComps()}} to extract results.
Also uses \code{\link[wtsUtilities:printGGList]{wtsUtilities::printGGList()}}.
}
