% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.ZScores.Biomass.R
\name{compareFits.ZScores.Biomass}
\alias{compareFits.ZScores.Biomass}
\title{Compare z-scores from fitting biomass time series by fleet among several model runs}
\usage{
compareFits.ZScores.Biomass(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}

\item{fleets}{- names of fleets to include (or "all")}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained', 'discard', or 'total')}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare z-scores from fitting biomass time series by fleet among
several model runs.
}
\details{
Uses [rTCSAM2013::getMDFR.SurveyQuantities()],
[rTCSAM2013::getMDFR.FisheryQuantities()], [rTCSAM02::getMDFR.ZScores.ABData()].
Also uses [wtsUtilities::printGGList].
}
