% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.ZScores.PrNatZ.R
\name{compareFits.ZScores.PrNatZ}
\alias{compareFits.ZScores.PrNatZ}
\title{Compare Pearson's residuals or nll residuals from size comps by fleet among several model runs}
\usage{
compareFits.ZScores.PrNatZ(
  objs = NULL,
  mdfr = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "total"),
  residuals.type = c("pearsons", "nlls"),
  tcsam2013.type = c("PRs_yxmz", "PRs_yxz"),
  BRT = 4,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}

\item{mdfr}{\itemize{
\item dataframe from call to \code{extractFits.ZScores.PrNatZ}  (as alternative to objs)
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all" or NULL to include all)
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained',or 'total')
}}

\item{residuals.type}{\itemize{
\item residual type for tcsam02 models ('pearsons' or 'nlls')
}}

\item{tcsam2013.type}{\itemize{
\item pearsons residuals type for tcsam2013 models ("PRs_yxmz","PRs_yxz")
}}

\item{BRT}{\itemize{
\item bad residual threshold (threshold to flag bad residuals, if residuals.type is 'pearsons')
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare Pearson's residuals or nll residuals from size comps by fleet among
several model runs.
}
\details{
Uses \code{\link[=extractFits.ZScores.PrNatZ]{extractFits.ZScores.PrNatZ()}} to extract data for plotting.
Also uses \code{\link[wtsUtilities:printGGList]{wtsUtilities::printGGList()}}.
}
