% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePopProcesses.R
\name{comparePopProcesses}
\alias{comparePopProcesses}
\title{Compare population processes from TCSAM2013 and 2015 model runs}
\usage{
comparePopProcesses(reps2013, reps2015)
}
\arguments{
\item{reps2013}{- TCSAM2013 report file object (or named list of objects)}

\item{reps2015}{- TCSAM2015 report file object (or named list of objects)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population processes from TCSAM2013 and 2015 model runs.
}

