% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePopQuants.R
\name{comparePopQuants}
\alias{comparePopQuants}
\title{Compare population quantities from TCSAM2013 and 2015 model runs}
\usage{
comparePopQuants(reps2013 = NULL, reps2015 = NULL, rsims = NULL,
  showPlot = TRUE, pdf = NULL, width = 8, height = 6, verbose = FALSE)
}
\arguments{
\item{reps2013}{- TCSAM2013 report file object (or named list of objects)}

\item{reps2015}{- TCSAM2015 report file object (or named list of objects)}

\item{rsims}{- rsimTCSAM results object (named list of objects)}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population quantities from TCSAM2013 and 2015 model runs.
}
\details{
None.
}

