% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Fisheries.Catchability.R
\name{compareResults.Fisheries.Catchability}
\alias{compareResults.Fisheries.Catchability}
\title{Function to plot fishery catchabilities by year using ggplot2}
\usage{
compareResults.Fisheries.Catchability(objs, fisheries = "all",
  years = "all", cast = "x", dodge = 0.2, facet_grid = "x~.",
  scales = "free_y", pdf = NULL, showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fisheries}{- vector of fisheries to plot, or "all"}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{dodge}{- width to dodge overlapping series}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{scales}{- parameter passed to ggplot2::facet_grid()}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
lists ofggplot2 objects, nested by fishery or an empty list if year is NULL
}
\description{
This function plots fishery catchability estimates by year,
  sex and maturity state.
}
\details{
None.
}
