% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Fisheries.CatchBiomass.R
\name{compareResults.Fisheries.CatchBiomass}
\alias{compareResults.Fisheries.CatchBiomass}
\title{Function to compare estimated fishery catch biomass by year among several models}
\usage{
compareResults.Fisheries.CatchBiomass(objs, category = c("captured",
  "discarded", "retained", "discard mortality", "index"), cast = NULL,
  facet_grid = NULL, facet_wrap = NULL, dodge = 0.2, years = "all",
  mxy = 15, nrow = 5, showPlot = TRUE, pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{category}{- 'captured','discarded','retained','discard mortality', or 'index'}

\item{cast}{- cast'ing formula for aggregating by factors (x,m,s,z)}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- 'all' or vector of years to include}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares estimated fishery catch biomass by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}

