% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.ManagementQuantities.R
\name{compareResults.ManagementQuantities}
\alias{compareResults.ManagementQuantities}
\title{Compare management quantities among model scenarios}
\usage{
compareResults.ManagementQuantities(objs, base = 1, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item single model resLst object, or named list of them
}}

\item{base}{\itemize{
\item name or index of base case for difference calculations
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
a list of ggplot2 plot objects
}
\description{
Function to compare management quantities among model scenarios.
}
\details{
Uses \code{\link[rTCSAM02:getMDFR.ManagementQuantities]{rTCSAM02::getMDFR.ManagementQuantities()}} to get the management quantities calculated
in each model case. Uses \code{\link[=computeDiffs]{computeDiffs()}} with \code{sense=1} to compute differences with base case.
}
