% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.ParameterValues.R
\name{compareResults.ParameterValues}
\alias{compareResults.ParameterValues}
\title{Function to compare parameter values among several models (UNDER DEVELOPMENT!)}
\usage{
compareResults.ParameterValues(
  objs,
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object(?)
}
\description{
This function compares parameter values
among several models.
}
\details{
Under development!
}
