% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Abundance.R
\name{compareResults.Pop.Abundance}
\alias{compareResults.Pop.Abundance}
\title{Function to compare population abundance estimates by year among several models}
\usage{
compareResults.Pop.Abundance(objs, type = c("N_yxms", "N_yxm", "N_yx"),
  numRecent = 15, dodge = 0.2, showPlot = TRUE, pdf = NULL,
  verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{type}{- type of abuundance ("N_yxms","N_yxm","N_yx","iN_xmsz","fN_xmsz")}

\item{numRecent}{- number of "recent" years to plot}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares population abundance estimates by year
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.PopQuantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY}.
}

