% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Abundance1.R
\name{compareResults.Pop.Abundance1}
\alias{compareResults.Pop.Abundance1}
\title{Function to compare population abundance estimates by year among several models}
\usage{
compareResults.Pop.Abundance1(objs, type = c("N_yxmsz", "N_yxmz", "N_yxz",
  "N_yxms", "N_yxm", "N_yx"), numRecent = 15, facet_grid = NULL,
  facet_wrap = NULL, dodge = 0.2, years = "all", mxy = 15, nrow = 5,
  showPlot = TRUE, pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{type}{- type of abuundance ("N_yxmsz","N_yxmz","N_yxz","N_yxms","N_yxm","N_yx")}

\item{numRecent}{- number of "recent" years to plot}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares population abundance estimates (aggregated or as size comps, depending on "type") by year
  among several models.
}
\details{
This function compares population abundance estimates (aggregated or as size comps, depending on "type") by year
  among several models. It uses \code{rTCSAM2013::getMDFR.Pop.Quantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, and to extract model results, and \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY} to plot them. The level of aggregation is based on the value for "type" (unlike 
\code{compreResults.Pop.Abundance}, where a cast'ing formula is specified.)
}

