% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Biomass.R
\name{compareResults.Pop.Biomass}
\alias{compareResults.Pop.Biomass}
\title{Function to compare biomass estimates by year among several models}
\usage{
compareResults.Pop.Biomass(objs, type = c("B_yxms", "B_yxm", "B_yx"),
  numRecent = 15, dodge = 0.2, showPlot = TRUE, pdf = NULL,
  verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{numRecent}{- number of "recent" years to plot}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares biomass estimates by year
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.PopQuantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY}.
}

