% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Biomass1.R
\name{compareResults.Pop.Biomass1}
\alias{compareResults.Pop.Biomass1}
\title{Function to compare biomass estimates by year among several models}
\usage{
compareResults.Pop.Biomass1(objs, type = c("B_yxmsz", "B_yxmz", "B_yxz",
  "B_yxms", "B_yxm", "B_yx"), numRecent = 15, facet_grid = NULL,
  facet_wrap = NULL, dodge = 0.2, years = "all", mxy = 15, nrow = 5,
  showPlot = TRUE, pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{type}{- type of biomass ("B_yxmsz","B_yxmz","B_yxz","B_yxms","B_yxm","B_yx")}

\item{numRecent}{- number of "recent" years to plot}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares biomass estimates by year
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.PopQuantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY}.
}

