% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.GrowthMatrices.LinePlots.R
\name{compareResults.Pop.GrowthMatrices.LinePlots}
\alias{compareResults.Pop.GrowthMatrices.LinePlots}
\title{Compare growth matrices among several model runs using line plots}
\usage{
compareResults.Pop.GrowthMatrices.LinePlots(objs, dodge = 0.2,
  ncol = 3, nrow = 5, showPlot = FALSE, pdf = NULL,
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{dodge}{- width to dodge overlapping series}

\item{ncol}{- number of columns per page for multi-year plots}

\item{nrow}{- number of rows per page (nominal) for multi-year plots}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list of ggplot objects, by sex
}
\description{
Function to compare growth matrices among several model runs using line plots.
}
\details{
The line plots are faceted by pre-molt size. If multiple models are compared, then sex-specific plots are
created, with individual lines corresponding to different models. If a single model is plotted, the
different sexes are plotted using different colors on the same faceted plots.
}
