% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.GrowthMatrices.R
\name{compareResults.Pop.GrowthMatrices}
\alias{compareResults.Pop.GrowthMatrices}
\title{Compare growth matrices among several model runs using bubble plots}
\usage{
compareResults.Pop.GrowthMatrices(objs, showPlot = FALSE, pdf = NULL,
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list with a ggplot object
}
\description{
Function to compare growth matrices among several model runs using bubble plots.
}
\details{
If multiple models are compared, then a set of bubble plots
are created faceted by mode case (row) and sex (column).
If a single model is plotted, then bubble plots are created faceted by sex (row). One ggplot
object is returned as a named list, with the figure caption as the name.
}
