% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.IF.NatZ.R
\name{compareResults.Pop.IF.NatZ}
\alias{compareResults.Pop.IF.NatZ}
\title{Function to compare initial/final population abundance estimates among several models}
\usage{
compareResults.Pop.IF.NatZ(
  objs,
  type = c("iN_xmsz", "fN_xmsz"),
  dodge = 0.2,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{type}{\itemize{
\item type of abuundance ("iN_xmsz","fN_xmsz")
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function compares initial/final population abundance estimates
among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.PopQuantities},
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and
\code{plotMDFR.XY}.
}
