% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.MeanGrowthPlusProbs.R
\name{compareResults.Pop.MeanGrowthPlusProbs}
\alias{compareResults.Pop.MeanGrowthPlusProbs}
\title{Compare mean and probability of post-molt size among several model runs}
\usage{
compareResults.Pop.MeanGrowthPlusProbs(
  objs,
  scale = 10,
  zbnds = c(25, 185),
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{scale}{\itemize{
\item scaling factor for probabilities
}}

\item{zbnds}{\itemize{
\item 2-element vector indicating lower and upper bounds on model size bins
}}

\item{showPlot}{\itemize{
\item flag to print plot to current device
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list of ggplot objects (plots by sex)
}
\description{
Function to compare mean and probability of post-molt size among several model runs.
}
\details{
None.
}
