% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.NaturalMortality.BarPlot.R
\name{compareResults.Pop.NaturalMortality.BarPlot}
\alias{compareResults.Pop.NaturalMortality.BarPlot}
\title{Function to compare natural mortality rates as bars}
\usage{
compareResults.Pop.NaturalMortality.BarPlot(
  objs,
  colour_scale = ggplot2::scale_color_hue(),
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects from TCSAM02
}}

\item{colour_scale}{\itemize{
\item ggplot2 scale_colour object (default is \code{\link[ggplot2:scale_hue]{ggplot2::scale_colour_hue()}})
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object as list element
}
\description{
This function compares natural mortality estimates by
sex and maturity state using barplots.
}
\details{
None.
}
