% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.NaturalMortality.R
\name{compareResults.Pop.NaturalMortality}
\alias{compareResults.Pop.NaturalMortality}
\title{Function to plot natural mortality rates by year using ggplot2}
\usage{
compareResults.Pop.NaturalMortality(objs, dodge = 0.2, showPlot = TRUE,
  pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function plots natural mortality estimates by year
  and sex for immatures, new shell matures and old shell matures.
}
\details{
None.
}

