% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.PrM2M.R
\name{compareResults.Pop.PrM2M}
\alias{compareResults.Pop.PrM2M}
\title{Compare probability of molt-to-maturity among several model runs}
\usage{
compareResults.Pop.PrM2M(
  objs,
  dodge = 0.2,
  colour_scale = ggplot2::scale_color_hue(),
  showPlot = TRUE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{colour_scale}{\itemize{
\item ggplot2 scale_colour object (default is \code{\link[ggplot2:scale_hue]{ggplot2::scale_colour_hue()}})
}}

\item{showPlot}{\itemize{
\item flag to print plot to current device
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot object
}
\description{
Function to compare probability of molt-to-maturity among several model runs.
}
\details{
None.
}
