% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Quantities.R
\name{compareResults.Pop.Quantities}
\alias{compareResults.Pop.Quantities}
\title{Compare population quantities from TCSAM2013, TCSAM02, and rsimTCSAM model runs}
\usage{
compareResults.Pop.Quantities(
  objs,
  plotPoints = FALSE,
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = TRUE,
  pdf = NULL,
  width = 8,
  height = 6,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item (named list of objects)
}}

\item{plotPoints}{\itemize{
\item flag to include points (default: FALSE)
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag to show/print plots immediately
}}

\item{pdf}{\itemize{
\item name of pdf file to record plot output to
}}

\item{width}{\itemize{
\item pdf page width (in inches)
}}

\item{height}{\itemize{
\item pdf page width (in inches)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debug info
}}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population quantities from TCSAM2013, TCSAM02, and rsimTCSAM model runs.
}
\details{
None.
}
