% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Availability.R
\name{compareResults.Surveys.Availability}
\alias{compareResults.Surveys.Availability}
\title{Function to plot survey availabilities by year using ggplot2}
\usage{
compareResults.Surveys.Availability(
  objs,
  fleets = "all",
  years = "all",
  cast = "x",
  dodge = 0.2,
  facet_grid = "x~.",
  scales = "free_y",
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects or dataframe from call to \code{extractMDFR.Surveys.Availability}
}}

\item{fleets}{\itemize{
\item vector of names of surveys to plot, or "all"
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{cast}{\itemize{
\item formula to exclude factors from "averaging" over
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{facet_grid}{\itemize{
\item ggplot2 formula to produce figure with gridded facets
}}

\item{scales}{\itemize{
\item parameter passed to ggplot2::facet_grid()
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list of ggplot2 objects
}
\description{
This function plots survey availability estimates by year,
sex and maturity state.
}
\details{
None.
}
