% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.AvlFcns.R
\name{compareResults.Surveys.AvlFcns}
\alias{compareResults.Surveys.AvlFcns}
\title{Function to compare survey availability functions by year among several models}
\usage{
compareResults.Surveys.AvlFcns(
  objs,
  cast = "y+x",
  fleets = "all",
  years = "all",
  dodge = 0.2,
  singlePlot = FALSE,
  mxy = 15,
  facet_wrap = NULL,
  facet_grid = ifelse(singlePlot, "x~case", "y~x"),
  plotPoints = FALSE,
  colour_scale = NULL,
  fill_scale = NULL,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects or dataframe from call to \code{extractMDFR.Surveys.AvlFcns}
}}

\item{cast}{\itemize{
\item formula to exclude factors from "averaging" over
}}

\item{fleets}{\itemize{
\item vector of fleets to plot, or "all"
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{singlePlot}{\itemize{
\item flag to plot all years on single plot (be sure to adjust facet_grid)
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{facet_wrap}{\itemize{
\item ggplot2 formula to produce figure with wrapped facets
}}

\item{facet_grid}{\itemize{
\item ggplot2 formula to produce figure with gridded facets
}}

\item{plotPoints}{\itemize{
\item flag to include points (default: FALSE)
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object, or null
}
\description{
This function compares survey availability functions by year
among several models.
}
\details{
None.
}
