% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Biomass.R
\name{compareResults.Surveys.Biomass}
\alias{compareResults.Surveys.Biomass}
\title{Function to compare estimated survey biomass by year among several models}
\usage{
compareResults.Surveys.Biomass(objs, category = "index", cast = "y+x",
  years = "all", facet_grid = "x~.", facet_wrap = NULL,
  scales = "fixed", dodge = 0.2, mxy = 15, nrow = 5, showPlot = FALSE,
  pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- cast'ing formula for aggregating by factors (y,x,m,s,z)}

\item{years}{- 'all' or vector of years to include}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{scales}{- parameter passed to ggplot2::facet_grid()}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares estimated survey biomass by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}

