% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Catchability.R
\name{compareResults.Surveys.Catchability}
\alias{compareResults.Surveys.Catchability}
\title{Function to plot survey catchabilities by year using ggplot2}
\usage{
compareResults.Surveys.Catchability(objs, cast = "x", years = "all",
  dodge = 0.2, facet_grid = "fleet~x", pdf = NULL, showPlot = FALSE,
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{dodge}{- width to dodge overlapping series}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function plots survey catchability estimates by year,
  sex and maturity state.
}
\details{
None.
}

