% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLatexForTable.ParamValues.Numbers.R
\name{createLatexForTable.ParamValues.Numbers}
\alias{createLatexForTable.ParamValues.Numbers}
\title{Function to print a table of parameters from several models}
\usage{
createLatexForTable.ParamValues.Numbers(
  mdfr,
  params = "all",
  fontsize = c("normal", "small"),
  table_type = c("", "longtable"),
  caption = "",
  label = "",
  conn = NULL,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item dataframe from call to \code{\link[=extractMDFR.Results.ParameterValues]{extractMDFR.Results.ParameterValues()}}, filtered for
}}

\item{params}{\itemize{
\item selected parameter names (default = "all")
}}

\item{fontsize}{\itemize{
\item font size (can be 'small')
}}

\item{table_type}{\itemize{
\item latex table type ("" or "longtable")
}}

\item{caption}{\itemize{
\item table caption
}}

\item{label}{\itemize{
\item table label (for cross-referencing)
}}

\item{conn}{\itemize{
\item NULL, "", a filename, or a connection object
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
latex text string representing the table (for later insertion into a markdown document)
}
\description{
This function prints a table of parameters from several models.
}
\details{
\code{mdfr} should be a dataframe obtained from \code{\link[=extractMDFR.Results.ParameterValues]{extractMDFR.Results.ParameterValues()}},
filtered to include only param_init_numbers and/or param_init_bounded_numbers.
The output format for values and standard deviations is 3 significant figures.

NOTE: \code{\link[wtsMarkdown:writeTabularToLatex]{wtsMarkdown::writeTabularToLatex()}} is used to create the latex code representing
the tabular object. If \code{conn=""}, the latex code is written to standard output. If \code{conn} is a filename or a
connection object, the latex code is written to that. If it is NULL, the code is not written to an output device.
In all cases, the code is returned as a string, invisibly.
}
