% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{createTable.ParamVals}
\alias{createTable.ParamVals}
\title{Create a table for parameter values for a given set of categories, processes, labels, and names}
\usage{
createTable.ParamVals(dfr, ctg = NULL, prc = NULL, devs = FALSE)
}
\arguments{
\item{dfr}{- dataframe with parameters at bounds}
}
\value{
a package{tables} tabular object
}
\description{
Function to create a package{tables} tabular object for parameter values for a
given set of categories, processes, labels, and names.
}
\details{
This function creates a tabular object parameter values for
a given set of parameter categories, processes, labels, and names.
}
