% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTables.R
\name{createTable.PsAtBs}
\alias{createTable.PsAtBs}
\title{Create a table for parameters at bounds}
\usage{
createTable.PsAtBs(dfr)
}
\arguments{
\item{dfr}{- dataframe with parameters at bounds}
}
\value{
a package{tables} tabular object
}
\description{
Function to create a package{tables} tabular object for parameters at bound.
}
\details{
This function creates a tabular object
}
