% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAggregatedDataLevels.R
\name{dropAggregatedDataLevels}
\alias{dropAggregatedDataLevels}
\title{Drop levels of sex, maturity state, and shell condition that have been aggregated in a model data dataframe}
\usage{
dropAggregatedDataLevels(tbl)
}
\arguments{
\item{tbl}{\itemize{
\item the dataframe to drop aggregated levels from
}}
}
\value{
an object of same class as tbl with aggregated levels dropped, leaving only the original unaggregated levels
}
\description{
Function to drop the levels of sex, maturity state, and shell condition that have been aggregated
in a canonical model data dataframe returned by, e.g., \code{\link[=extractMDFR.Data.FleetTimeSeriesABs]{extractMDFR.Data.FleetTimeSeriesABs()}}.
}
\details{
The model data dataframe should be in canconical format as returned by, e.g.,  \code{\link[=extractMDFR.Data.FleetTimeSeriesABs]{extractMDFR.Data.FleetTimeSeriesABs()}}.
The original levels in the input data to the associated TCSAM02 model are returned.
}
