% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFits.AllScores.ACD.R
\name{extractFits.AllScores.ACD}
\alias{extractFits.AllScores.ACD}
\title{Extract all scores from fits to biomass/abundance time series by fleet among several model runs as a dataframe}
\usage{
extractFits.AllScores.ACD(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "total"),
  data.type = c("biomass", "abundance"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam02.resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all" or NULL to include all)
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained',or 'total')
}}

\item{data.type}{\itemize{
\item 'biomass' or 'abundance'
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe in canonical format.
}
\description{
Function to extract all scores from biomass/abundance time series by fleet among
several model runs.
}
\details{
Uses \code{\link[rTCSAM02:getMDFR.AllScores.Biomass]{rTCSAM02::getMDFR.AllScores.Biomass()}} or \code{\link[rTCSAM02:getMDFR.AllScores.Abundance]{rTCSAM02::getMDFR.AllScores.Abundance()}}.
}
