% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFits.ZScores.ACD.R
\name{extractFits.ZScores.ACD}
\alias{extractFits.ZScores.ACD}
\title{Extract residuals from fits to biomass/abundance time series by fleet among several model runs as a dataframe}
\usage{
extractFits.ZScores.ACD(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "total"),
  data.type = c("biomass", "abundance"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{- object that can be converted into a list of tcsam02.resLst objects}

\item{fleets}{- names of fleets to include (or "all" or NULL to include all)}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',or 'total')}

\item{data.type}{- 'biomass' or 'abundance'}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format(?).
}
\description{
Function to extract residuals from biomass/abundance time series by fleet among
several model runs.
}
\details{
Uses [rTCSAM02::getMDFR.ZScores.Biomass()] or [rTCSAM02::getMDFR.ZScores.Abundance()].
Also uses [wtsUtilities::printGGList()].
}
