% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Data.FleetTimeSeriesABs.R
\name{extractMDFR.Data.FleetTimeSeriesABs}
\alias{extractMDFR.Data.FleetTimeSeriesABs}
\title{Extract time series of abundance or biommass data by fleet among several model scenarios}
\usage{
extractMDFR.Data.FleetTimeSeriesABs(
  objs = NULL,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  data.type = c("abundance", "biomass"),
  fleets = "all",
  sexs = "all",
  maturity_states = "all",
  shell_conditions = "all",
  ci = 0.8,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained', 'discard' or 'total')
}}

\item{data.type}{\itemize{
\item data type ('abundance' or 'biomass')
}}

\item{fleets}{\itemize{
\item vector of names of fleets to plot (or 'all')
}}

\item{sexs}{\itemize{
\item vector of sexes to plot (or 'all')
}}

\item{maturity_states}{\itemize{
\item vector of maturity states to plot (or 'all')
}}

\item{shell_conditions}{\itemize{
\item vector of shell conditions to plot (or 'all')
}}

\item{ci}{\itemize{
\item confidence intervals
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe
}
\description{
Function to extract abundance or biomass time series data by fleet among
several model scenarios.
}
\details{
Uses \code{\link[rTCSAM02:getMDFR.Data.FleetTimeSeries]{rTCSAM02::getMDFR.Data.FleetTimeSeries()}}.
}
