% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.ExploitationRates.R
\name{extractMDFR.ExploitationRates}
\alias{extractMDFR.ExploitationRates}
\title{Function to extract exploitation rates by year among several models}
\usage{
extractMDFR.ExploitationRates(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts exploitation rates by year
  among several models.
}
\details{
Results are extracted using \code{rTCSAM2013::getMDFR.Pop.Biomass},
\code{rsimTCSAM::getMDFR.Pop.Biomass}, and/or \code{rTCSAM02::getMDFR.Pop.Biomass}, as appropriate, and
cast to aggregate. This differs from \code{compareResults.Pop.Biomass1}.
}
