% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.Catchability.R
\name{extractMDFR.Fisheries.Catchability}
\alias{extractMDFR.Fisheries.Catchability}
\title{Function to extract fishery catchabilities by year using ggplot2}
\usage{
extractMDFR.Fisheries.Catchability(objs, cast = "x", verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
This function extracts fishery catchability estimates by year,
  sex and maturity state.
}
\details{
None.
}

