% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.CatchAbundance.R
\name{extractMDFR.Fisheries.CatchAbundance}
\alias{extractMDFR.Fisheries.CatchAbundance}
\title{Function to extract estimated fishery catch abundance by year among several models as a dataframe}
\usage{
extractMDFR.Fisheries.CatchAbundance(
  objs,
  fleets = "all",
  category = c("captured", "discarded", "retained", "discard mortality", "index"),
  cast = NULL,
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{category}{\itemize{
\item 'captured','discarded','retained','discard mortality', or 'index'
}}

\item{cast}{\itemize{
\item cast'ing formula for aggregating by factors (y,x,m,s,z)
}}

\item{years}{\itemize{
\item 'all' or vector of years to include
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe in canonical format
}
\description{
This function extracts estimated fishery catch abundance by year
among several models as a dataframe.
}
\details{
None.
}
