% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.RetFcns.R
\name{extractMDFR.Fisheries.RetFcns}
\alias{extractMDFR.Fisheries.RetFcns}
\title{Function to extract fishery retention functions by year among several models}
\usage{
extractMDFR.Fisheries.RetFcns(
  objs,
  fleets = "all",
  cast = "y+x",
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{cast}{\itemize{
\item formula to exclude factors from "averaging" over
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe in canonical format
}
\description{
This function extracts fishery retention functions by year
among several models.
}
\details{
None.
}
