% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.SelFcns.R
\name{extractMDFR.Fisheries.SelFcns}
\alias{extractMDFR.Fisheries.SelFcns}
\title{Function to compare fishery selectivity functions by year among several models}
\usage{
extractMDFR.Fisheries.SelFcns(
  objs,
  fleets = "all",
  cast = "y+x",
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fleets}{- names of fleets to include (or "all")}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares fishery selectivity functions by year
  among several models.
}
\details{
None.
}
