% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.AbundanceData.R
\name{extractMDFR.Fits.AbundanceData}
\alias{extractMDFR.Fits.AbundanceData}
\title{Extract fits to abundance time series by fleet among several model runs}
\usage{
extractMDFR.Fits.AbundanceData(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  ci = 0.8,
  fishery.pdfType = c("norm2", "normal", "lognormal"),
  verbose = FALSE
)
}
\arguments{
\item{fleets}{- names of fleets to include (or "all")}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{ci}{- confidence interval for plots}

\item{fishery.pdfType}{- assumed distribution for fishery data}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
dataframe
}
\description{
Function to extract fits to abundance time series by fleet among
several model runs.
}
\details{
Uses [rTCSAM2013::getMDFR.SurveyQuantities()],
[rTCSAM2013::getMDFR.FisheryQuantities()], [rTCSAM02::getMDFR.Fits.FleetData()].
}
