% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.EffectiveNs.R
\name{extractMDFR.Fits.EffectiveNs}
\alias{extractMDFR.Fits.EffectiveNs}
\title{Function to extract effective Ns from size comps by year among several models}
\usage{
extractMDFR.Fits.EffectiveNs(
  objs,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  category = c("index", "captured", "discarded", "retained", "discard mortality"),
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all" or NULL to include all)
}}

\item{fleet.type}{\itemize{
\item 'survey','fishery'
}}

\item{category}{\itemize{
\item 'total','discard','retained','discard mortality', or 'index'
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe
}
\description{
This function extracts effective Ns from size comps by year
among several models.
}
\details{
Results are extracted using \code{rTCSAM02::getMDFR.Fits.EffectiveNs}, as appropriate.
}
