% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.GrowthData.R
\name{extractMDFR.Fits.GrowthData}
\alias{extractMDFR.Fits.GrowthData}
\title{Extract fits to growth data among several model runs as a dataframe}
\usage{
extractMDFR.Fits.GrowthData(objs)
}
\arguments{
\item{objs}{- list of resLst objects}
}
\value{
dataframe
}
\description{
Function to extract fits to growth data among several model runs as a dataframe.
}
\details{
None.
}
