% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.MaturityOgiveData.R
\name{extractMDFR.Fits.MaturityOgiveData}
\alias{extractMDFR.Fits.MaturityOgiveData}
\title{Extract fits to maturity ogive data among several model runs as a dataframe}
\usage{
extractMDFR.Fits.MaturityOgiveData(objs)
}
\arguments{
\item{objs}{\itemize{
\item list of tcsam02 resLst objects
}}
}
\value{
dataframe
}
\description{
Function to extract fits to maturity ogive data among several model runs as a dataframe.
}
\details{
None.
}
