% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.Biomass.R
\name{extractMDFR.Pop.Biomass}
\alias{extractMDFR.Pop.Biomass}
\title{Function to extract estimated population biomass by year among several models as a dataframe}
\usage{
extractMDFR.Pop.Biomass(objs, cast = "y+x", years = "all", verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- casting formula for excluding y,x,m,s,z factor levels from sums across the unspecified factors}

\item{years}{- 'all' or vector of years to include}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts estimated population biomass by year
  among several models as a dataframe.
}
\details{
Results are extracted using \code{rTCSAM2013::getMDFR.Pop.Biomass},
\code{rsimTCSAM::getMDFR.Pop.Biomass}, and/or \code{rTCSAM02::getMDFR.Pop.Biomass}, as appropriate, and
cast to aggregate. This differs from \code{extractMDFR.Pop.Biomass1}.
}
