% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.GrowthMatrices.R
\name{extractMDFR.Pop.GrowthMatrices}
\alias{extractMDFR.Pop.GrowthMatrices}
\title{Extract growth matrices among several model runs}
\usage{
extractMDFR.Pop.GrowthMatrices(objs, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
Function to extract growth matrices among several model runs.
}
\details{
None.
}

