% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.MatureBiomass.R
\name{extractMDFR.Pop.MatureBiomass}
\alias{extractMDFR.Pop.MatureBiomass}
\title{Function to extract mature biomass estimates by year among several models}
\usage{
extractMDFR.Pop.MatureBiomass(objs, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
datafram in canonical format
}
\description{
This function extracts mature biomass estimates by year
among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.Pop.Quantities},
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rTCSAM02::getMDFR.Pop.Quantities}.
}
