% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.MeanGrowth.R
\name{extractMDFR.Pop.MeanGrowth}
\alias{extractMDFR.Pop.MeanGrowth}
\title{Extract mean post-molt size from several model runs}
\usage{
extractMDFR.Pop.MeanGrowth(objs, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in caonical format
}
\description{
Function to extract mean post-molt size from several model runs.
}
\details{
None.
}

