% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.NaturalMortality.R
\name{extractMDFR.Pop.NaturalMortality}
\alias{extractMDFR.Pop.NaturalMortality}
\title{Function to plot natural mortality rates by year using ggplot2}
\usage{
extractMDFR.Pop.NaturalMortality(objs, type = "M_yxm", verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{type}{\itemize{
\item type of estimates to return
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function plots natural mortality estimates by year,
sex and maturity state.
}
\details{
None.
}
