% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.PrM2M.R
\name{extractMDFR.Pop.PrM2M}
\alias{extractMDFR.Pop.PrM2M}
\title{Extract probability of molt-to-maturity from several model runs}
\usage{
extractMDFR.Pop.PrM2M(objs, verbose = FALSE)
}
\arguments{
\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- list of resLst objects}
}
\value{
dataframe in canonical format
}
\description{
Function to extract probability of molt-to-maturity from several model runs.
}
\details{
None.
}
