% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.Recruitment.R
\name{extractMDFR.Pop.Recruitment}
\alias{extractMDFR.Pop.Recruitment}
\title{Function to extract recruitment estimates by year among several models}
\usage{
extractMDFR.Pop.Recruitment(objs, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
This function extracts recruitment estimates by year
  among several models.
}
\details{
None.
}

