% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Pop.RecSizeDistribution.R
\name{extractMDFR.Pop.RecSizeDistribution}
\alias{extractMDFR.Pop.RecSizeDistribution}
\title{Extract recruitment size distributions from several model runs}
\usage{
extractMDFR.Pop.RecSizeDistribution(objs, verbose = FALSE)
}
\arguments{
\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- list of resLst objects}
}
\value{
dataframe in canonical format
}
\description{
Function to extract recruitment size distributions from several model runs.
}
\details{
None.
}
