% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Results.ParametersAtBounds.R
\name{extractMDFR.Results.ParametersAtBounds}
\alias{extractMDFR.Results.ParametersAtBounds}
\title{Extract parameters at one of their bounds from several model runs as a dataframe}
\usage{
extractMDFR.Results.ParametersAtBounds(objs, delta = 1e-04,
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{delta}{- relative fraction of range defining "at bounds"}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts parameters at one of their bounds from several model runs as a dataframe.
}
\details{
Results are extracted using \code{rTCSAM02::getMDFR.ParametersAtBounds} for tcsam02 model runs.
}
