% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.SdRep.RecAndSSB.R
\name{extractMDFR.SdRep.RecAndSSB}
\alias{extractMDFR.SdRep.RecAndSSB}
\title{Extract sd_report info on recruitment and SSB time series estimates as a dataframe}
\usage{
extractMDFR.SdRep.RecAndSSB(objs, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{ci}{\itemize{
\item confidence interval
}}

\item{verbose}{\itemize{
\item flag to print debugging info
}}
}
\value{
a dataframe
}
\description{
Function to extract sd_report info on recruitment and SSB time series estimates as a dataframe.
}
\details{
Delegates to \code{\link[rTCSAM02:getMDFR.SdRep.RecAndSSB]{rTCSAM02::getMDFR.SdRep.RecAndSSB()}}. Recruitment values are on ln-scale.
}
