% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.AvlFcns.R
\name{extractMDFR.Surveys.AvlFcns}
\alias{extractMDFR.Surveys.AvlFcns}
\title{Function to extract survey availability functions by year among several models}
\usage{
extractMDFR.Surveys.AvlFcns(
  objs,
  fleets = "all",
  cast = "y+x",
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fleets}{- names of fleets to include (or "all")}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts survey availability functions by year
  among several models.
}
\details{
None.
}
