% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.Biomass.R
\name{extractMDFR.Surveys.Biomass}
\alias{extractMDFR.Surveys.Biomass}
\title{Function to compare estimated survey biomass by year among several models}
\usage{
extractMDFR.Surveys.Biomass(
  objs,
  fleets = "all",
  category = "index",
  cast = "y+x",
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{cast}{\itemize{
\item cast'ing formula for aggregating by factors (y,x,m,s,z)
}}

\item{years}{\itemize{
\item 'all' or vector of years to include
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe in canonical format.
}
\description{
This function extracts estimated survey biomass by year
among several models as a dataframe
}
\details{
None.
}
