% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.SelFcns.R
\name{extractMDFR.Surveys.SelFcns}
\alias{extractMDFR.Surveys.SelFcns}
\title{Function to extract survey selectivity functions by year among several models}
\usage{
extractMDFR.Surveys.SelFcns(
  objs,
  fleets = "all",
  cast = "y+x",
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of survey fleets to include (or "all" or NULL to include all)
}}

\item{cast}{\itemize{
\item formula to exclude factors from "averaging" over
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
dataframe
}
\description{
This function extracts survey selectivity functions by year
among several models.
}
\details{
None.
}
