% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.Pop.MatureBiomass.R
\name{extractResults.Pop.MatureBiomass}
\alias{extractResults.Pop.MatureBiomass}
\title{Function to extract mature biomass estimates by year among several models}
\usage{
extractResults.Pop.MatureBiomass(objs, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
datafram in canonical format
}
\description{
This function extracts mature biomass estimates by year
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.Pop.Quantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}.
}

