% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResLsts.R
\name{getResLsts}
\alias{getResLsts}
\title{Create a list of TCSAM2013, TCSAM02 and rsimTCSAM resLst objects from a set of model runs}
\usage{
getResLsts(inp.dirs = NULL)
}
\arguments{
\item{inp.dirs}{- a dataframe with columns "case", "modelType", "path", "repType", "modelName", "prsType"}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
a named list of the appropriate resLst objects.
}
\description{
Function to create a list of TCSAM2013, TCSAM02 and rsimTCSAM resLst objects from a set of model runs.
}
\details{
Uses \code{getResLst} to open a folders with model runs listed in inp.dirs.
}

