% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.Characteristics.Surveys.R
\name{modelComparisons.Characteristics.Surveys}
\alias{modelComparisons.Characteristics.Surveys}
\title{Render a document of comparison plots of estimated survey characteristics from several models}
\usage{
modelComparisons.Characteristics.Surveys(
  models,
  surveys = "all",
  selYears = c(1975, 1982, 1988),
  avlYears = NULL,
  capYears = c(1975, 1982, 1988),
  mxy = 5,
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.EstimatedCharacteristics.Surveys.Rmd", package
    = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{- named list of model results (as resLst objects) to compare}

\item{surveys}{- vector of names of surveys to plot, or "all"}

\item{selYears}{- vector of years for which to plot selectivity functions ("all"=all years; NULL=don't plot)}

\item{avlYears}{- vector of years for which to plot availability functions ("all"=all years; NULL=don't plot)}

\item{capYears}{- vector of years for which to plot capture probability functions ("all"=all years; NULL=don't plot)}

\item{mxy}{- number of functions to plot per page, when plotting selectivity, availability, or capture probability functions}

\item{output_format}{- "word_document" or "pdf_document"}

\item{output_dir}{- path to folder to use for output}

\item{rmd}{- Rmd file to process (defalut="modelComparisons.Rmd")}

\item{docx_styles}{- full path to Word (docx) style template for Word documents}

\item{pdf_styles}{- full path to style template for pdf documents}

\item{clean}{- T/F to delete intermediate files}

\item{rmd_dir}{- folder enclosing rmd file}
}
\description{
Function to render a document of comparison plots of estimated survey
characteristics from several models.
}
\details{
The default resulting document title will be of the form "ModelComparisons.EstimatedCharacteristics.Surveys.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
