% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.ModelFits.ACD.Surveys.R
\name{modelComparisons.ModelFits.ACD.Surveys}
\alias{modelComparisons.ModelFits.ACD.Surveys}
\title{Render a document of comparison plots for model fits to aggregated survey abundance/biomass data}
\usage{
modelComparisons.ModelFits.ACD.Surveys(models, plot1stObs = TRUE,
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.ModelFits.ACD.Surveys.Rmd",
  package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package =
  "wtsUtilities"), pdf_styles = system.file("rmd/StylesForRmdPDFs.sty",
  package = "wtsUtilities"), clean = FALSE)
}
\arguments{
\item{models}{- named list of model results (as resLst objects) to compare}

\item{plot1stObs}{- flag to plot observed data only from first model}

\item{output_format}{- "word_document" or "pdf_document"}

\item{output_dir}{- path to folder to use for output}

\item{rmd}{- Rmd file to process (defalut="rmd/modelComparisons.ModelFits.ACD.Surveys.Rmd")}

\item{docx_styles}{- full path to Word (docx) style template for Word documents}

\item{pdf_styles}{- full path to style template for pdf documents}

\item{clean}{- T/F to delete intermediate files}

\item{rmd_dir}{- folder enclosing rmd file}
}
\description{
Function to render a document of comparison plots for model fits to
aggregated survey abundance/biomass data.
}
\details{
Resulting document title will be of the form "ModelComparisons.ModelFits.ACD.Surveys.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
