% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.ModelFits.OtherData.R
\name{modelComparisons.ModelFits.OtherData}
\alias{modelComparisons.ModelFits.OtherData}
\title{Render a document of comparison plots for model fits to 'other' data}
\usage{
modelComparisons.ModelFits.OtherData(
  models,
  plot1stObs = TRUE,
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.ModelFits.OtherData.Rmd", package =
    "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{\itemize{
\item named list of model results (as resLst objects) to compare
}}

\item{plot1stObs}{\itemize{
\item flag to plot observed data only from first model
}}

\item{output_format}{\itemize{
\item "word_document" or "pdf_document"
}}

\item{output_dir}{\itemize{
\item path to folder to use for output
}}

\item{rmd}{\itemize{
\item Rmd file to process (defalut="modelComparisons.Rmd")
}}

\item{docx_styles}{\itemize{
\item full path to Word (docx) style template for Word documents
}}

\item{pdf_styles}{\itemize{
\item full path to style template for pdf documents
}}

\item{clean}{\itemize{
\item T/F to delete intermediate files
}}

\item{rmd_dir}{\itemize{
\item folder enclosing rmd file
}}
}
\description{
Function to render a document of comparison plots for model fits to
'other' data.
}
\details{
Resulting document title will be of the form "ModelComparisons.ModelFits.OtherData.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
