% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/modelComparisons.ModelFits.ZCsByYear.Fisheries.R
\name{modelComparisons.ModelFits.ZCsByYear.Fisheries}
\alias{modelComparisons.ModelFits.ZCsByYear.Fisheries}
\title{Render a document of comparison plots for model fits to fishery size composition data by year}
\usage{
modelComparisons.ModelFits.ZCsByYear.Fisheries(
  models,
  fleets = "all",
  years = "all",
  plotRetained = TRUE,
  plotTotal = TRUE,
  plot1stObs = TRUE,
  nrow = 5,
  ncol = 4,
  useBars = TRUE,
  usePins = FALSE,
  usePinsAndPts = FALSE,
  useLines = TRUE,
  usePoints = TRUE,
  pinSize = 0.2,
  lineSize = 1,
  pointSize = 1,
  alpha = 0.5,
  stripText = ggplot2::element_text(),
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.ModelFits.ZCsByYear.Fisheries.Rmd", package =
    "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{\itemize{
\item named list of model results (as resLst objects) to compare
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{years}{\itemize{
\item years to plot, as numerical vector (or "all" to plot all years)
}}

\item{plotRetained}{\itemize{
\item flag to plot retained catch size comps
}}

\item{plotTotal}{\itemize{
\item flag to plot total catch size comps
}}

\item{plot1stObs}{\itemize{
\item flag (T/F) to plot observations only from first case, or character vector cases cases from which to plot observations
}}

\item{nrow}{\itemize{
\item number of rows per page for output plots
}}

\item{ncol}{\itemize{
\item number of columns per page for output plots
}}

\item{useBars}{\itemize{
\item flag to use bars for observations
}}

\item{usePins}{\itemize{
\item flag to use pins for observations
}}

\item{usePinsAndPts}{\itemize{
\item flag to add pts to observations when pins are used
}}

\item{useLines}{\itemize{
\item flag to use lines for predictions
}}

\item{usePoints}{\itemize{
\item flag to use points for predictions
}}

\item{pinSize}{\itemize{
\item width of pin line
}}

\item{lineSize}{\itemize{
\item prediction line size
}}

\item{pointSize}{\itemize{
\item prediction point size
}}

\item{alpha}{\itemize{
\item prediction transparency
}}

\item{stripText}{\itemize{
\item \code{\link[ggplot2:element]{ggplot2::element_text()}} object describing font and margin to use for panel strips
}}

\item{output_format}{\itemize{
\item "word_document" or "pdf_document"
}}

\item{output_dir}{\itemize{
\item path to folder to use for output
}}

\item{rmd}{\itemize{
\item Rmd file to process (defalut="rmd/modelComparisons.ModelFits.ZCsByYear.Fisheries.Rmd")
}}

\item{docx_styles}{\itemize{
\item full path to Word (docx) style template for Word documents
}}

\item{pdf_styles}{\itemize{
\item full path to style template for pdf documents
}}

\item{clean}{\itemize{
\item T/F to delete intermediate files
}}

\item{rmd_dir}{\itemize{
\item folder enclosing rmd file
}}
}
\description{
Function to render a document of comparison plots for model fits to
fishery size composition data by year.
}
\details{
Resulting document title will be of the form "ModelComparisons.ModelFits.ZCsByYear.Fisheries.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
