% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.ModelFits.ZCsByYear.Surveys.R
\name{modelComparisons.ModelFits.ZCsByYear.Surveys}
\alias{modelComparisons.ModelFits.ZCsByYear.Surveys}
\title{Render a document of comparison plots for model fits to survey size composition data by year}
\usage{
modelComparisons.ModelFits.ZCsByYear.Surveys(
  models,
  fleets = "all",
  years = "all",
  plot1stObs = TRUE,
  nrow = 5,
  ncol = 4,
  useBars = TRUE,
  usePins = FALSE,
  usePinsAndPts = FALSE,
  useLines = TRUE,
  usePoints = TRUE,
  pinSize = 0.2,
  lineSize = 1,
  pointSize = 1,
  alpha = 0.5,
  stripText = ggplot2::element_text(),
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.ModelFits.ZCsByYear.Surveys.Rmd", package =
    "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{- named list of model results (as resLst objects) to compare}

\item{fleets}{- names of fleets to include (or "all")}

\item{years}{- years to plot, as numerical vector (or "all" to plot all years)}

\item{plot1stObs}{- flag (T/F) to plot observations only from first case, or character vector cases cases from which to plot observations}

\item{nrow}{- number of rows per page for output plots}

\item{ncol}{- number of columns per page for output plots}

\item{useBars}{- flag to use bars for observations}

\item{usePins}{- flag to use pins for observations}

\item{usePinsAndPts}{- flag to add pts to observations when pins are used}

\item{useLines}{- flag to use lines for predictions}

\item{usePoints}{- flag to use points for predictions}

\item{pinSize}{- width of pin line}

\item{lineSize}{- prediction line size}

\item{pointSize}{- prediction point size}

\item{alpha}{- prediction transparency}

\item{stripText}{- [ggplot2::element_text()] object describing font and margin to use for panel strips}

\item{output_format}{- "word_document" or "pdf_document"}

\item{output_dir}{- path to folder to use for output}

\item{rmd}{- Rmd file to process (defalut="rmd/modelComparisons.ModelFits.ZCsByYear.Surveys.Rmd")}

\item{docx_styles}{- full path to Word (docx) style template for Word documents}

\item{pdf_styles}{- full path to style template for pdf documents}

\item{clean}{- T/F to delete intermediate files}

\item{rmd_dir}{- folder enclosing rmd file}
}
\description{
Function to render a document of comparison plots for model fits to
survey size composition data by year.
}
\details{
Resulting document title will be of the form "ModelComparisons.ModelFits.ZCsByYear.Surveys.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
