% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.PopQuantities.R
\name{modelComparisons.PopQuantities}
\alias{modelComparisons.PopQuantities}
\title{Render a document of comparison plots of estimated population quantities from several models}
\usage{
modelComparisons.PopQuantities(
  models,
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.PopQuantities.Rmd", package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{\itemize{
\item named list of model results (as resLst objects) to compare
}}

\item{output_format}{\itemize{
\item "word_document" or "pdf_document"
}}

\item{output_dir}{\itemize{
\item path to folder to use for output
}}

\item{rmd}{\itemize{
\item Rmd file to process (defalut="modelComparisons.Rmd")
}}

\item{docx_styles}{\itemize{
\item full path to Word (docx) style template for Word documents
}}

\item{pdf_styles}{\itemize{
\item full path to style template for pdf documents
}}

\item{clean}{\itemize{
\item T/F to delete intermediate files
}}

\item{rmd_dir}{\itemize{
\item folder enclosing rmd file
}}
}
\description{
Function to render a document of comparison plots of estimated population
quantities from several models.
}
\details{
Resulting document title will be of the form "ModelComparisons.PopQuantities.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
