% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.MaturityOgiveData}
\alias{plotLollipops.MaturityOgiveData}
\title{plot lollipops for fits to maturity ogive data}
\usage{
plotLollipops.MaturityOgiveData(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 5,
  nrow = 5,
  xlab = "size (mm CW)",
  ylab = type
)
}
\arguments{
\item{mdfr}{- input dataframe from [rCompTCMs::extractMDFR.Fits.MaturityOgiveData()]}

\item{type}{- type of value to plot (e.g., "zscores")}

\item{extreme}{- criteria for extreme values}

\item{dw}{- dodge width (x-axis units)}

\item{nrow}{- number of rows for faceting}

\item{xlab}{- x-axis label}

\item{ylab}{- y-axis label}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to maturity ogive data.
}
\details{
Plots \code{type} as lollipops against size, faceted by year.
}
