% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.ZCs}
\alias{plotLollipops.ZCs}
\title{plot lollipops for fits to size composition data}
\usage{
plotLollipops.ZCs(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 5,
  size = 0.1,
  xlab = "size bin (mm CW)",
  ylab = "z-score",
  facet_wrap = ~y,
  ncol = 2
)
}
\arguments{
\item{mdfr}{- input data}

\item{type}{- type of value to plot (e.g., "zscores")}

\item{extreme}{- criteria for extreme values}

\item{dw}{- dodge width (x-axis units)}

\item{size}{- line size for lollipop stems}

\item{xlab}{- x-axis label}

\item{ylab}{- y-axis label}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to size composition data.
}
\details{
Under construction!
}
