% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.BarLine.R
\name{plotMDFR.BarLine}
\alias{plotMDFR.BarLine}
\title{Plot melted dataframe using a barchart/line-plot format.}
\usage{
plotMDFR.BarLine(
  mdfr,
  x = NULL,
  value.var = "val",
  agg.formula = NULL,
  agg.function = sum,
  ...,
  colour = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  faceting = NULL,
  xlab = "",
  ylab = "",
  units = "",
  lnscale = FALSE,
  title = "",
  guideTitleColour = NULL,
  guideTitleFill = NULL,
  guideTitleLineType = NULL,
  guideTitleShape = NULL,
  showPlot = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item melted dataframe
}}

\item{x}{\itemize{
\item column name for x-axis values
}}

\item{value.var}{\itemize{
\item column name for values to aggregate (value.var in cast)/plot on y-axis
}}

\item{agg.formula}{\itemize{
\item aggregation formula (left-hand side of cast formula)
}}

\item{agg.function}{\itemize{
\item aggregation function (fun.aggregate in cast)
}}

\item{...}{\itemize{
\item further arguments passed to aggregating function
}}

\item{colour}{\itemize{
\item column name to which colour aesthetic is mapped
}}

\item{fill}{\itemize{
\item column name to which fill aesthetic is mapped
}}

\item{linetype}{\itemize{
\item column name to which linetype aesthetic is mapped
}}

\item{shape}{\itemize{
\item column name to which shape aesthetic is mapped
}}

\item{faceting}{\itemize{
\item faceting formula
}}

\item{xlab}{\itemize{
\item x axis label
}}

\item{ylab}{\itemize{
\item y axis label
}}

\item{units}{\itemize{
\item combined with y axis label
}}

\item{lnscale}{\itemize{
\item flag to plot on ln-scale
}}

\item{title}{\itemize{
\item plot title
}}

\item{guideTitleColour}{\itemize{
\item title for colour guide
}}

\item{guideTitleFill}{\itemize{
\item title for fill guide
}}

\item{guideTitleLineType}{\itemize{
\item title for linetype guide
}}

\item{guideTitleShape}{\itemize{
\item title for shape guide
}}

\item{showPlot}{\itemize{
\item flag to show plot immediately
}}
}
\value{
ggplot2 object
}
\description{
Function to plot a melted dataframe using a barchart/line-plot format.
}
\details{
uses \code{reshape2} package.
}
