% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Fits.SizeComps.R
\name{plotMDFR.Fits.SizeComps}
\alias{plotMDFR.Fits.SizeComps}
\title{Plot fits to size comps by fleet among several model runs from a dataframe}
\usage{
plotMDFR.Fits.SizeComps(mdfr = NULL, ylab = "", cap1 = "1",
  plotObsAsBars = TRUE, nrow = 5, ncol = 3, pdf = NULL,
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfr}{- dataframe from call to \code{rCompTCMs::extractFits.SizeComps}}

\item{ylab}{- y axis label}

\item{cap1}{- template for caption}

\item{plotObsAsBars}{- flag (T/F) to plot observations/predictions as bars/lines (T) or lines/bars (F)}

\item{nrow}{- number of rows per page for output plots}

\item{ncol}{- number of columns per page for output plots}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to plot fits to size comps by fleet among 
several model runs from a dataframe.
}
\details{
Uses \code{wtsUtilities::printGGList}.
}

