% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Fits.TimeSeries.R
\name{plotMDFR.Fits.TimeSeries}
\alias{plotMDFR.Fits.TimeSeries}
\title{Plot comparisons of fits to time series from a set of model runs, with a "zoom" for recent years}
\usage{
plotMDFR.Fits.TimeSeries(dfr, plot1stObs = TRUE, numRecent = 15, x = "y",
  y = "val", lci = "lci", uci = "uci", case = "case", type = "type",
  facets = NULL, scales = "fixed", plotObs = TRUE, plotMod = TRUE,
  xlab = "year", ylab = NULL, title = NULL, xlims = NULL,
  ylims = NULL, showPlot = FALSE)
}
\arguments{
\item{dfr}{- dataframe}

\item{plot1stObs}{- flag to plot observations from first case only}

\item{numRecent}{- number of "recent" years to plot}

\item{x}{- column name with x axis values}

\item{y}{- column name with y axis values}

\item{lci}{- column name with y axis values}

\item{uci}{- column name with y axis values}

\item{case}{- column name with case names}

\item{type}{- column name with type values (i.e., "observed","predicted")}

\item{facets}{- string giving faceting formula}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{plotObs}{- plot observations}

\item{plotMod}{- plot model fits/predictions}

\item{xlab}{-}

\item{ylab}{-}

\item{title}{-}

\item{xlims}{-}

\item{ylims}{-}

\item{showPlot}{-}
}
\value{
list with requested ggplot objects
}
\description{
Function to plot a comparison of fits to time series from a set of model runs, with a "zoom" for recent years.
}
\details{
numRecent provides the "zoom" for a second set of faceted plots including
only the most recent years. Calls \code{plotMDFR.Fits.TimeSeries1}.
}

