% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Fits.TimeSeries1.R
\name{plotMDFR.Fits.TimeSeries1}
\alias{plotMDFR.Fits.TimeSeries1}
\title{Plot comparisons of fits to time series from a set of model runs}
\usage{
plotMDFR.Fits.TimeSeries1(
  dfr,
  plot1stObs = TRUE,
  x = "y",
  y = "val",
  lci = "lci",
  uci = "uci",
  case = "case",
  type = "type",
  facets = NULL,
  position = position_dodge(0.2),
  scales = "fixed",
  plotObs = TRUE,
  plotMod = TRUE,
  xlab = "year",
  ylab = NULL,
  title = NULL,
  xlims = NULL,
  ylims = NULL,
  colour_scale = ggplot2::scale_color_hue(),
  fill_scale = ggplot2::scale_fill_hue(),
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe
}}

\item{plot1stObs}{\itemize{
\item flag to plot observations from first case only
}}

\item{x}{\itemize{
\item column name with x axis values
}}

\item{y}{\itemize{
\item column name with y axis values
}}

\item{lci}{\itemize{
\item column name with y axis values
}}

\item{uci}{\itemize{
\item column name with y axis values
}}

\item{case}{\itemize{
\item column name with case names
}}

\item{type}{\itemize{
\item column name with type values (i.e., "observed","predicted")
}}

\item{facets}{\itemize{
\item grid faceting formula (as an expresssion for multilevel faceting)
}}

\item{position}{\itemize{
\item indicates ggplot2 position_ to use ('dodge','jitter','identity',)
}}

\item{scales}{\itemize{
\item ggplot2 scales option for facet_grid
}}

\item{plotObs}{\itemize{
\item plot observations
}}

\item{plotMod}{\itemize{
\item plot case fits/predictions
}}

\item{xlab}{\itemize{
\item 
}}

\item{ylab}{\itemize{
\item 
}}

\item{title}{\itemize{
\item 
}}

\item{xlims}{\itemize{
\item 
}}

\item{ylims}{\itemize{
\item 
}}

\item{colour_scale}{\itemize{
\item ggplot2 scale_colour object (default is \code{\link[ggplot2:scale_hue]{ggplot2::scale_colour_hue()}})
}}

\item{fill_scale}{\itemize{
\item ggplot2 scale_fill object (default is \code{\link[ggplot2:scale_hue]{ggplot2::scale_fill_hue()}})
}}

\item{showPlot}{\itemize{
\item flag (T/F) to print plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print debugging info
}}

\item{ci}{\itemize{
\item confidence interval for error bars
}}

\item{pdfType}{\itemize{
\item assumed error distribution for confidence intervals
}}
}
\value{
ggplot object
}
\description{
Function to plot a comparison of fits to time series from a set of model runs.
}
\details{
\code{facets} should be given as an expression, not as a character string, if you want multilevel faceting.
}
