% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.XY.R
\name{plotMDFR.XY}
\alias{plotMDFR.XY}
\title{Plot melted dataframe using an xy-plot format.}
\usage{
plotMDFR.XY(
  mdfr,
  x = NULL,
  value.var = "val",
  agg.formula = NULL,
  agg.function = sum,
  ...,
  colour = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  dodge = 0,
  facet_grid = NULL,
  facet_wrap = NULL,
  nrow = NULL,
  ncol = NULL,
  dir = "h",
  scales = "fixed",
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  ylab = "",
  units = "",
  lnscale = FALSE,
  title = "",
  guideTitleColour = NULL,
  guideTitleFill = NULL,
  guideTitleLineType = NULL,
  guideTitleShape = NULL,
  plotPoints = TRUE,
  plotLines = TRUE,
  plotABline = FALSE,
  abline = list(intercept = 0, slope = 1, colour = "black", linetype = 3, size = 1, alpha
    = 0.8),
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item melted dataframe
}}

\item{x}{\itemize{
\item column name for x-axis values
}}

\item{value.var}{\itemize{
\item column name for values to aggregate (value.var in cast)/plot on y-axis
}}

\item{agg.formula}{\itemize{
\item aggregation formula (left-hand side of cast formula)
}}

\item{agg.function}{\itemize{
\item aggregation function (fun.aggregate in cast)
}}

\item{...}{\itemize{
\item further arguments passed to aggregating function
}}

\item{colour}{\itemize{
\item column name to which colour aesthetic is mapped
}}

\item{fill}{\itemize{
\item column nameto which fill aesthetic is mapped
}}

\item{linetype}{\itemize{
\item column name to which linetype aesthetic is mapped
}}

\item{shape}{\itemize{
\item column name to which shape aesthetic is mapped
}}

\item{dodge}{\itemize{
\item width to dodge overlapping points/lines
}}

\item{facet_grid}{\itemize{
\item faceting formula for facet_grid'ing
}}

\item{facet_wrap}{\itemize{
\item faceting formula for facet_wrap'ing
}}

\item{nrow}{\itemize{
\item number of rows when using facet_wrap
}}

\item{ncol}{\itemize{
\item number of columns when using facet_wrap
}}

\item{dir}{\itemize{
\item fastest direction ('h','v') for faceting
}}

\item{scales}{\itemize{
\item scaling flag when faceting
}}

\item{xlim}{\itemize{
\item x axis limits
}}

\item{ylim}{\itemize{
\item y axis limits
}}

\item{xlab}{\itemize{
\item x axis label
}}

\item{ylab}{\itemize{
\item y axis label
}}

\item{units}{\itemize{
\item combined with y axis label
}}

\item{lnscale}{\itemize{
\item flag to plot on ln-scale
}}

\item{title}{\itemize{
\item plot title
}}

\item{guideTitleColour}{\itemize{
\item title for colour guide
}}

\item{guideTitleFill}{\itemize{
\item title for fill guide
}}

\item{guideTitleLineType}{\itemize{
\item title for linetype guide
}}

\item{guideTitleShape}{\itemize{
\item title for shape guide
}}

\item{plotPoints}{\itemize{
\item flag to plot points
}}

\item{plotLines}{\itemize{
\item flag to plot lines
}}

\item{plotABline}{\itemize{
\item flag to plot a straight line
}}

\item{abline}{\itemize{
\item list w/ components intercept, slope, colour, size, linetype, alpha describing line to plots
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag to show plot immediately
}}
}
\value{
ggplot2 object
}
\description{
Function to plot a melted dataframe using an xy plot format.
}
\details{
uses \code{reshape2} package.
}
