% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.CohortProgression.R
\name{plotPop.CohortProgression}
\alias{plotPop.CohortProgression}
\title{Function to plot cohort progression by year among several models}
\usage{
plotPop.CohortProgression(
  mdfr,
  facet_grid = NULL,
  facet_wrap = NULL,
  dodge = 0.2,
  years = NULL,
  mxy = 4,
  nrow = 2,
  lnscale = FALSE,
  scales = "fixed",
  types = c("progression", "byyear"),
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{- melted data.frame with cohort progressions}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- vector of years to include in plots (or NULL for all)}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{lnscale}{- use log scale on y axis}

\item{scales}{- scales parameter for facet_grid/facet_wrap}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
a list of ggplot2 objects
}
\description{
This function plots cohort progression by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}
