% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.MeanGrowthPlusProbs.R
\name{plotPop.MeanGrowthPlusProbs}
\alias{plotPop.MeanGrowthPlusProbs}
\title{Plot mean and probability of post-molt size for several model runs}
\usage{
plotPop.MeanGrowthPlusProbs(
  mdfrMnG,
  mdfrPrG,
  scale = 10,
  xbnds = NULL,
  ybnds = NULL,
  zbnds = c(25, 185),
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfrMnG}{\itemize{
\item melted dataframe for mean growth (from calling \code{rTCSAM02::getMDFR.Pop.MeanGrowth})
}}

\item{mdfrPrG}{\itemize{
\item melted dataframe with growth probabilities (from calling \code{rTCSAM02::getMDFR.Pop.GrowthMatrices})
}}

\item{scale}{\itemize{
\item scaling factor for probabilities
}}

\item{zbnds}{\itemize{
\item 2-element vector indicating lower and upper bounds on model size bins
}}

\item{showPlot}{\itemize{
\item flag to print plot to current device
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list of ggplots by sex
}
\description{
Function to plot mean and probability of post-molt size for several model runs.
}
\details{
Mean post-molt size as a function of pre-molt size is plotted as a line for each case (model),
overlaid on the (scaled) actual probability of post-molt size.
Cases are distinguished by fill and line colours.
A 1:1 line and lines indicating the model size limits are also shown.
Plots are produced by sex, and a list of plots is returned.
}
