% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.NaturalMortality.R
\name{plotPop.NaturalMortality}
\alias{plotPop.NaturalMortality}
\title{Function to plot natural mortality rates by year using ggplot2}
\usage{
plotPop.NaturalMortality(
  mdfr,
  dodge = 0.2,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item melted dataframe
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object as list element
}
\description{
This function plots natural mortality estimates by year,
sex and maturity state.
}
\details{
None.
}
