% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.RecSizeDistribution.R
\name{plotPop.RecSizeDistribution}
\alias{plotPop.RecSizeDistribution}
\title{Plot recruitment size distributions among several model runs}
\usage{
plotPop.RecSizeDistribution(
  mdfr,
  dodge = 0.2,
  xbnds = NULL,
  plotPoints = plotPoints,
  colour_scale = colour_scale,
  fill_scale = fill_scale,
  showPlot = TRUE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item melted dataframe
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{xbnds}{\itemize{
\item two-element vector with x-axis bound (or NULL)
}}

\item{plotPoints}{\itemize{
\item flag to include points on plot
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag to print plot to current device
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot object
}
\description{
Function to plot recruitment size distributions among several model runs.
}
\details{
None.
}
