% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.ParametersAtBounds.R
\name{printResults.ParametersAtBounds}
\alias{printResults.ParametersAtBounds}
\title{Function to print a table of parameters at bounds for several models}
\usage{
printResults.ParametersAtBounds(objs, delta = 1e-04, caption = "",
  format = c("latex", "docx"), landscape = FALSE, font = "11pt",
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{showPlot}{- flag (T/F) to show plot}
}
\value{
ggplot2 object
}
\description{
This function prints a table of parameters at bounds for several models.
}
\details{
Uses \code{extractMDFR.Results.ParametersAtBounds} to get a dataframe of parameters at one of their bounds,
based upon the value for delta. If format="latex", the package \code{tables} is used to create and print a table
in latex format.
}
