% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.ParameterValues.R
\name{printResults.ParameterValues}
\alias{printResults.ParameterValues}
\title{Function to print a table of parameters from several models}
\usage{
printResults.ParameterValues(
  objs = NULL,
  mdfr = NULL,
  categories = "all",
  processes = "all",
  types = "all",
  params = "all",
  format = c("latex", "docx"),
  landscape = FALSE,
  fontsize = "11pt",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{mdfr}{\itemize{
\item dataframe from call to \code{\link[=extractMDFR.Results.ParameterValues]{extractMDFR.Results.ParameterValues()}} (as alternative to objs)
}}

\item{categories}{\itemize{
\item default = "all"
}}

\item{processes}{\itemize{
\item default = "all"
}}

\item{types}{\itemize{
\item default = "all"
}}

\item{params}{\itemize{
\item selected parameter names (default = "all")
}}

\item{format}{\itemize{
\item 'latex' or 'docx'
}}

\item{landscape}{\itemize{
\item flag indicating whether page orientation is landscape or portrait
}}

\item{fontsize}{\itemize{
\item font size (can be 'small')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function prints a table of parameters from several models.
}
\details{
Uses \code{\link[=extractMDFR.Results.ParameterValues]{extractMDFR.Results.ParameterValues()}} to get a dataframe of parameters.
If format="latex", the package \pkg{tables} is used to create and print a table
in latex format. Output format for values and standard deviations is 3 significant figures.

Specific tables can be created/selected by specifying non-default values for categories,
processes, types, and params.
}
