% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroFunctions.R
\name{retroCalcMohnsRho}
\alias{retroCalcMohnsRho}
\title{Calculate Mohn's rho from a dataframe of retrospective estimates}
\usage{
retroCalcMohnsRho(dfr, ylab = "value", yadj = 0, showPlot = FALSE)
}
\arguments{
\item{dfr}{- dataframe of retrospective estimates for analysis}

\item{ylab}{- y axis label for plot}

\item{yadj}{- adjustment to \code{assessment year-1} for max year in dataframe}

\item{showPlot}{- flag (T/F) to show plot immediately}
}
\value{
Returns a list with rho, a dataframe of values, and a ggplot2 object, invisibly
}
\description{
This function calculate Mohn's rho from a dataframe of retrospective estimates,
and optionally produces a plot.
}
\details{
Mohn's rho quantifies the relative bias exhibited by estimates of some
quantity from a set of retrospective model runs.
}
