% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroFunctions.R
\name{retroGetMDFR}
\alias{retroGetMDFR}
\title{Get a melted dataframe of model estimates from a list of retrospective model results}
\usage{
retroGetMDFR(modsRetro, FUN, yadj = 0, ...)
}
\arguments{
\item{modsRetro}{\itemize{
\item named list of resLst objects, with names corresponding to assessment years representing peels
}}

\item{FUN}{\itemize{
\item rCompTCMs function used to extract the desired estimates (i.e., an extractMDFR function)
}}

\item{yadj}{\itemize{
\item adjustment to extract correct max year in results, relative to \code{assessment year-1}
}}

\item{...}{\itemize{
\item additional arguments passed to FUN
}}
}
\value{
dataframe
}
\description{
This function extracts a melted dataframe from a list of retrospective model results.
}
\details{
Results are extracted using \code{FUN} for a retrospective series of tcsam02 model runs.
}
\note{
The names of the resLst objects should correspond to the \emph{assessment year} they represent.
\code{yadj} should be 0 (the default) for results that end in the "max year" of the assessment,
which is one less than the assessment year (e.g., recruitment, mature biomass). For results that end
with the year of the assessment (e.g., survey biomass, numbers-at size), \code{yadj} should be 1.
}
