% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.FleetTimeSeriesABs.R
\name{compareData.FleetTimeSeriesABs}
\alias{compareData.FleetTimeSeriesABs}
\title{Compare time series of abundance or biommass data by fleet among several model scenarios}
\usage{
compareData.FleetTimeSeriesABs(
  objs = NULL,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  data.type = c("abundance", "biomass"),
  fleets = "all",
  sexs = "all",
  maturity_states = "all",
  shell_conditions = "all",
  position = ggplot2::position_dodge(0.2),
  ci = 0.8,
  numRecent = 15,
  ylims = NULL,
  facets = "fleet~x",
  scales = "free_y",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects (or dataframe from call to \code{\link[=extractMDFR.Data.FleetTimeSeriesABs]{extractMDFR.Data.FleetTimeSeriesABs()}})
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained', 'discard' or 'total')
}}

\item{data.type}{\itemize{
\item data type ('abundance' or 'biomass')
}}

\item{fleets}{\itemize{
\item vector of names of fleets to plot (or 'all')
}}

\item{sexs}{\itemize{
\item vector of sexes to plot (or 'all')
}}

\item{maturity_states}{\itemize{
\item vector of maturity states to plot (or 'all')
}}

\item{shell_conditions}{\itemize{
\item vector of shell conditions to plot (or 'all')
}}

\item{position}{\itemize{
\item indicates ggplot2 position_ to use ('dodge','jitter','identity',)
}}

\item{ci}{\itemize{
\item confidence interval for plots
}}

\item{numRecent}{\itemize{
\item number of years for 'recent' plot
}}

\item{ylims}{\itemize{
\item limits for y axis (default is NULL)
}}

\item{facets}{\itemize{
\item grid faceting formula
}}

\item{scales}{\itemize{
\item ggplot2 scales option for facet_grid
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare abundance or biomass time series data by fleet among
several model scenarios.
}
\details{
Uses \code{\link[=extractMDFR.Data.FleetTimeSeriesABs]{extractMDFR.Data.FleetTimeSeriesABs()}} to extract the data (if objs is not a dataframe).
Uses \code{\link[=plotMDFR.Fits.TimeSeries]{plotMDFR.Fits.TimeSeries()}} to make the plots, with \code{plotObs=TRUE} and \code{plotMod=FALSE}. Only time series with
at least one non-zero value will be plotted (to eliminate lots of values along y=0).
}
