% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDiffs.R
\name{compareDiffs}
\alias{compareDiffs}
\title{Function to compare differences of a quantity among several models}
\usage{
compareDiffs(dfr, base = 1, diff.type = c("percent", "absolute"),
  diff.min = 0.001, cast = NULL, facet_grid = NULL, facet_wrap = NULL,
  scales = "fixed", dodge = 0.2, title = "", years = "all",
  showPlot = FALSE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{dfr}{- dataframe in canonical format}

\item{diff.type}{- "percent" or "absolute"}

\item{diff.min}{- minimum difference to show}

\item{cast}{- cast'ing formula for aggregating by factors before differencing (fleet,x,m,s,z,y)}

\item{facet_grid}{- formula (or string version of formula) for faceting using facet_grid}

\item{facet_wrap}{- one-sided formula (e.g., "~y+x") or character vector (e.g., c('y','x')) for faceting using facet_wrap}

\item{scales}{- scales parameter for use with facet_grid/facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- 'all' or vector of years to include}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}
}
\value{
list of ggplot2 objects
}
\description{
This function compares differences of a quantity
  among several models.
}
\details{
If 'z' and 'y' are cast'ing factors, then a set of annual size composition plots are produced. 
If 'z' is a cast'ing factor but not 'y', 'z' will be used as the x axis. Otherwise,
a set of time series plots are produced.
}
