% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.BiomassData.R
\name{compareFits.BiomassData}
\alias{compareFits.BiomassData}
\title{Compare fits to biomass time series by fleet among several model runs}
\usage{
compareFits.BiomassData(objs = NULL, fleet.type = c("survey",
  "fishery"), catch.type = c("index", "retained", "discard", "total"),
  ci = 0.8, fishery.pdfType = c("norm2", "normal", "lognormal"),
  numRecent = 15, plot1stObs = TRUE, scales = "free_y", pdf = NULL,
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{ci}{- confidence interval for plots}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects, OR a melted dataframe}
}
\value{
Non-nested list of ggplot2 objects, with captions as names, or NULL.
}
\description{
Function to compare fits to biomass time series by fleet among
several model runs.
}
\details{
If \code{objs} is a list, then \code{extractMDFR.Fits.BiomassData} is used to extract a melted dataframe
in canonical format with the biomass observations and predictions to be plotted. \code{objs} can also be a melted
dataframe from a previous call to \code{extractMDFR.Fits.BiomassData}.
Also uses \code{wtsUtilities::printGGList}.
}
