% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.EffectiveNs.R
\name{compareFits.EffectiveNs}
\alias{compareFits.EffectiveNs}
\title{Function to compare effective Ns from size compositions by year among several models}
\usage{
compareFits.EffectiveNs(objs = NULL, mdfr = NULL,
  fleet.type = c("survey", "fishery"), category = c("index",
  "captured", "discarded", "retained", "discard mortality"),
  years = "all", dodge = 0.2, facet_wrap = NULL,
  facet_grid = "x+m+s~.", showPlot = FALSE, pdf = NULL,
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fleet.type}{- 'survey','fishery'}

\item{category}{- 'total','discard','retained','discard mortality', or 'index'}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{dodge}{- width to dodge overlapping series}

\item{facet_wrap}{- ggplot2 formula to produce figure with wrapped facets}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
lists ofggplot2 objects, nested by fishery
}
\description{
This function compares effective Ns from size compositions by year
  among several models.
}
\details{
None.
}
