% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.EffortData.R
\name{compareFits.EffortData}
\alias{compareFits.EffortData}
\title{Compare fits to effort data among several model runs}
\usage{
compareFits.EffortData(objs, dodge = 0.2, pdf = NULL,
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{dodge}{- width to dodge overlapping series}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag to print plot to current device}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot object
}
\description{
Function to compare fits to effort data among several model runs.
}
\details{
None.
}
