% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.GrowthData.R
\name{compareFits.GrowthData}
\alias{compareFits.GrowthData}
\title{Compare fits to growth data among several model runs}
\usage{
compareFits.GrowthData(objs, dodge = 0.2, plot1stObs = FALSE, pdf = NULL,
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{dodge}{- width to dodge overlapping series}

\item{plot1stObs}{- flag to include observations only from 1st model scenario}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag to print plot to current device}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot object
}
\description{
Function to compare fits to growth data among several model runs.
}
\details{
None.
}
