% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.MeanSizeComps.R
\name{compareFits.MeanSizeComps}
\alias{compareFits.MeanSizeComps}
\title{Compare fits to mean size comps by fleet among several model runs}
\usage{
compareFits.MeanSizeComps(
  objs = NULL,
  mdfr = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  years = "all",
  ci = 0.8,
  plot1stObs = TRUE,
  facet_grid = "m+s~x",
  scales = "free_y",
  linesize = 1,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}

\item{mdfr}{- melted dataframe from call to \code{extractFits.MeanSizeComps} (as alternative to objs)}

\item{fleets}{- names of fleets to include (or "all")}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{years}{- years to plot, as numerical vector (or "all" to plot all years)}

\item{ci}{- confidence interval}

\item{plot1stObs}{- flag (T/F) to plot observations only from first case}

\item{facet_grid}{- faceting formula for ggplot2::facet_grid()}

\item{scales}{- scales parameter for ggplot2::facet_grid()}

\item{linesize}{- line width (in mm)}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list of ggplot2 objects, with captions as names
}
\description{
Function to compare fits to mean size comps by fleet among
several model runs.
}
\details{
Uses \code{rTCSAM2013::getMDFR.SurveyQuantities()},
\code{rTCSAM2013::getMDFR.FisheryQuantities()}, \code{rTCSAM02::getMDFR.Fits.FleetData()}.
Also uses \code{reshape2::dcast}, \code{wtsUtilities::calcCIs}, \code{wtsUtilities::printGGList}.
}
