% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.SizeComps.R
\name{compareFits.SizeComps}
\alias{compareFits.SizeComps}
\title{Compare fits to size comps by fleet among several model runs}
\usage{
compareFits.SizeComps(objs = NULL, fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"), nrow = 5,
  ncol = 3, numRecent = 15, plot1stObs = TRUE, scales = "free_y",
  pdf = NULL, showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{nrow}{- number of rows per page for output plots}

\item{ncol}{- number of columns per page for output plots}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
Non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare fits to size comps by fleet among 
several model runs.
}
\details{
Uses \code{rTCSAM2013::getMDFR.SurveyQuantities()},
\code{rTCSAM2013::getMDFR.FisheryQuantities()}, \code{rTCSAM02::getMDFR.Fits.FleetData()}.
Also uses \code{wtsUtilities::printGGList}.
}

