% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.ZScores.Biomass.R
\name{compareFits.ZScores.Biomass}
\alias{compareFits.ZScores.Biomass}
\title{Compare z-scores from fitting biomass time series by fleet among several model runs}
\usage{
compareFits.ZScores.Biomass(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained', 'discard', or 'total')
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare z-scores from fitting biomass time series by fleet among
several model runs.
}
\details{
Uses \code{\link[rTCSAM2013:getMDFR.SurveyQuantities]{rTCSAM2013::getMDFR.SurveyQuantities()}},
\code{\link[rTCSAM2013:getMDFR.FisheryQuantities]{rTCSAM2013::getMDFR.FisheryQuantities()}}, \code{\link[rTCSAM02:getMDFR.ZScores.ABData]{rTCSAM02::getMDFR.ZScores.ABData()}}.
Also uses \link[wtsUtilities:printGGList]{wtsUtilities::printGGList}.
}
