% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFits.ZScores.PrNatZ.R
\name{compareFits.ZScores.PrNatZ}
\alias{compareFits.ZScores.PrNatZ}
\title{Compare Pearson's residuals or nll residuals from size comps by fleet among several model runs}
\usage{
compareFits.ZScores.PrNatZ(objs = NULL, fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "total"),
  residuals.type = c("pearsons", "nlls"), tcsam2013.type = c("PRs_yxmz",
  "PRs_yxz"), showPlot = FALSE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',or 'total')}

\item{residuals.type}{- residual type for tcsam02 models ('pearsons' or 'nlls')}

\item{tcsam2013.type}{- pearsons residuals type for tcsam2013 models ("PRs_yxmz","PRs_yxz")}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
non-nested list of ggplot2 objects, with captions as names
}
\description{
Function to compare Pearson's residuals or nll residuals from size comps by fleet among 
several model runs.
}
\details{
Uses \code{rTCSAM2013::getMDFR.SurveyQuantities()},
\code{rTCSAM2013::getMDFR.FisheryQuantities()}, \code{rTCSAM02::getMDFR.ZScores.PrNatZ()}.
Also uses \code{wtsUtilities::printGGList}.
}

