% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePopQuants.R
\name{comparePopQuants}
\alias{comparePopQuants}
\title{Compare population quantities from TCSAM2013 and TCSAM02 model runs}
\usage{
comparePopQuants(objs, showPlot = TRUE, pdf = NULL, width = 8,
  height = 6, verbose = FALSE)
}
\arguments{
\item{objs}{-  (named list of objects)}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population quantities from TCSAM2013 and TCSAM02 model runs.
}
\details{
None.
}

