% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Fisheries.Catchability.R
\name{compareResults.Fisheries.Catchability}
\alias{compareResults.Fisheries.Catchability}
\title{Function to plot fishery catchabilities by year using ggplot2}
\usage{
compareResults.Fisheries.Catchability(
  objs,
  fleets = "all",
  years = "all",
  cast = "x",
  dodge = 0.2,
  facet_grid = "x~.",
  scales = "free_y",
  plotPoints = TRUE,
  colour_scale = NULL,
  fill_scale = NULL,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item vector of fisheries to plot, or "all"
}}

\item{years}{\itemize{
\item vector of years to show, or 'all' to show all years
}}

\item{cast}{\itemize{
\item formula to exclude factors from "averaging" over
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{facet_grid}{\itemize{
\item ggplot2 formula to produce figure with gridded facets
}}

\item{scales}{\itemize{
\item parameter passed to ggplot2::facet_grid()
}}

\item{plotPoints}{\itemize{
\item flag to plot points
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
lists ofggplot2 objects, nested by fishery or an empty list if year is NULL
}
\description{
This function plots fishery catchability estimates by year,
sex and maturity state.
}
\details{
None.
}
