% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Fisheries.CatchAbundance.R
\name{compareResults.Fisheries.CatchAbundance}
\alias{compareResults.Fisheries.CatchAbundance}
\title{Function to compare estimated fishery catch abundance by year among several models}
\usage{
compareResults.Fisheries.CatchAbundance(
  objs,
  fleets = "all",
  category = c("captured", "discarded", "retained", "discard mortality", "index"),
  cast = NULL,
  facet_grid = NULL,
  facet_wrap = NULL,
  scales = "fixed",
  dodge = 0.2,
  years = "all",
  mxy = 15,
  nrow = 5,
  lnscale = FALSE,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fleets}{- names of fleets to include (or "all")}

\item{category}{- 'captured','discarded','retained','discard mortality', or 'index'}

\item{cast}{- cast'ing formula for aggregating by factors (x,m,s,z)}

\item{facet_grid}{- formula (or string version of formula) for faceting using facet_grid}

\item{facet_wrap}{- one-sided formula (e.g., "~y+x") or character vector (e.g., c('y','x')) for faceting using facet_wrap}

\item{scales}{- scales parameter for use with facet_grid/facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- 'all' or vector of years to include}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{lnscale}{- use log scale on y axis}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list of ggplot2 objects
}
\description{
This function compares estimated fishery catch abundance by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}
