% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Fisheries.CatchBiomass.R
\name{compareResults.Fisheries.CatchBiomass}
\alias{compareResults.Fisheries.CatchBiomass}
\title{Function to compare estimated fishery catch biomass by year among several models}
\usage{
compareResults.Fisheries.CatchBiomass(
  objs,
  fleets = "all",
  category = c("captured", "discarded", "retained", "discard mortality", "index"),
  cast = NULL,
  facet_grid = NULL,
  facet_wrap = NULL,
  scales = "free_y",
  dodge = 0.2,
  years = "all",
  mxy = 15,
  nrow = 5,
  plotPoints = TRUE,
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{category}{\itemize{
\item 'captured','discarded','retained','discard mortality', or 'index'
}}

\item{cast}{\itemize{
\item cast'ing formula for aggregating by factors (x,m,s,z)
}}

\item{facet_grid}{\itemize{
\item formula (or string version of formula) for faceting using facet_grid
}}

\item{facet_wrap}{\itemize{
\item one-sided formula (e.g., "~y+x") or character vector (e.g., c('y','x')) for faceting using facet_wrap
}}

\item{scales}{\itemize{
\item scales parameter for use with facet_grid/facet_wrap
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{years}{\itemize{
\item 'all' or vector of years to include
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{nrow}{\itemize{
\item number of rows per page, when facet_wrap'ing
}}

\item{plotPoints}{\itemize{
\item flag to plot points
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list of ggplot2 plot objects
}
\description{
This function compares estimated fishery catch biomass by year
among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}
