% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.ParameterValues.R
\name{compareResults.ParameterValues}
\alias{compareResults.ParameterValues}
\title{Function to compare parameter values among several models (UNDER DEVELOPMENT!)}
\usage{
compareResults.ParameterValues(objs, showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object(?)
}
\description{
This function compares parameter values
among several models.
}
\details{
Under development!
}
