% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Abundance.R
\name{compareResults.Pop.Abundance}
\alias{compareResults.Pop.Abundance}
\title{Function to compare estimated population abundance by year among several models}
\usage{
compareResults.Pop.Abundance(objs, cast = "x", facet_grid = NULL,
  facet_wrap = NULL, dodge = 0.2, years = "all", mxy = 15,
  nrow = 5, lnscale = FALSE, scales = "fixed", showPlot = FALSE,
  pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- casting formula for excluding x,m,s,z factor levels from sums across the unspecified factors}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- 'all' or vector of years to include}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{lnscale}{- use log scale on y axis}

\item{scales}{- scales parameter for facet_grid/facet_wrap}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares estimated population abundance by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced. Results are extracted using \code{rTCSAM2013::getMDFR.Pop.Abundance},
\code{rsimTCSAM::getMDFR.Pop.Abundance}, and/or \code{rTCSAM02::getMDFR.Pop.Abundance}, as appropriate, and 
cast to aggregate. This differs from \code{compareResults.Pop.Abundance1}.
}
