% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Abundance1.R
\name{compareResults.Pop.Abundance1}
\alias{compareResults.Pop.Abundance1}
\title{Function to compare population abundance estimates by year among several models}
\usage{
compareResults.Pop.Abundance1(
  objs,
  type = c("N_yxmsz", "N_yxmz", "N_yxz", "N_yxms", "N_yxm", "N_yx"),
  years = "all",
  numRecent = 15,
  facet_grid = NULL,
  facet_wrap = NULL,
  scales = "fixed",
  dodge = 0.2,
  mxy = 15,
  nrow = 5,
  plotPoints = TRUE,
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{type}{\itemize{
\item type of abuundance ("N_yxmsz","N_yxmz","N_yxz","N_yxms","N_yxm","N_yx")
}}

\item{years}{\itemize{
\item "all" or numerical vector of years to include
}}

\item{numRecent}{\itemize{
\item number of "recent" years to plot
}}

\item{facet_grid}{\itemize{
\item formula for faceting using facet_grid
}}

\item{facet_wrap}{\itemize{
\item formula for faceting using facet_wrap
}}

\item{scales}{\itemize{
\item ggplot2 scales option for facet_grid
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{nrow}{\itemize{
\item number of rows per page, when facet_wrap'ing
}}

\item{plotPoints}{\itemize{
\item flag to include points
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function compares population abundance estimates (aggregated or as size comps, depending on "type") by year
among several models.
}
\details{
This function compares population abundance estimates (aggregated or as size comps, depending on "type") by year
among several models. It uses \code{rTCSAM2013::getMDFR.Pop.Quantities},
\code{rsimTCSAM::getMDFR.Pop.Quantities}, and to extract model results, and \code{rsimTCSAM::getMDFR.Pop.Quantities}, and
\code{plotMDFR.XY} to plot them. The level of aggregation is based on the value for "type" (unlike
\code{compreResults.Pop.Abundance}, where a cast'ing formula is specified.)
}
