% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Biomass.R
\name{compareResults.Pop.Biomass}
\alias{compareResults.Pop.Biomass}
\title{Function to compare estimated population biomass by year among several models}
\usage{
compareResults.Pop.Biomass(
  objs,
  cast = "x",
  facet_grid = "x~.",
  facet_wrap = NULL,
  dodge = 0.2,
  years = "all",
  mxy = 15,
  nrow = 5,
  lnscale = FALSE,
  scales = "fixed",
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{cast}{\itemize{
\item casting formula for excluding x,m,s,z factor levels from sums across the unspecified factors
}}

\item{facet_grid}{\itemize{
\item formula for faceting using facet_grid
}}

\item{facet_wrap}{\itemize{
\item formula for faceting using facet_wrap
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{years}{\itemize{
\item 'all' or vector of years to include
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{nrow}{\itemize{
\item number of rows per page, when facet_wrap'ing
}}

\item{lnscale}{\itemize{
\item use log scale on y axis
}}

\item{scales}{\itemize{
\item scales parameter for facet_grid/facet_wrap
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function compares estimated population biomass by year
among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced. Results are extracted using \code{\link[rTCSAM2013:getMDFR.Pop.Biomass]{rTCSAM2013::getMDFR.Pop.Biomass()}},
\code{\link[rsimTCSAM:getMDFR.Pop.Biomass]{rsimTCSAM::getMDFR.Pop.Biomass()}}, and/or [rTCSAM02::getMDFR.Pop.Biomass()], as appropriate, and
cast to aggregate. This differs from \code{\link[=compareResults.Pop.Biomass1]{compareResults.Pop.Biomass1()}}.
}
