% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Biomass1.R
\name{compareResults.Pop.Biomass1}
\alias{compareResults.Pop.Biomass1}
\title{Function to compare population biomass estimates by year among several models}
\usage{
compareResults.Pop.Biomass1(objs, type = c("B_yxmsz", "B_yxmz", "B_yxz",
  "B_yxms", "B_yxm", "B_yx"), years = "all", numRecent = 15,
  facet_grid = NULL, facet_wrap = NULL, scales = "fixed", dodge = 0.2,
  mxy = 15, nrow = 5, showPlot = FALSE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{type}{- type of biomass ("B_yxmsz","B_yxmz","B_yxz","B_yxms","B_yxm","B_yx")}

\item{years}{- "all" or numerical vector of years to include}

\item{numRecent}{- number of "recent" years to plot}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{scales}{- ggplot2 scales option for facet_grid}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares biomass estimates (aggregated or as size comps, depending on "type") by year
  among several models.
}
\details{
This function compares biomass estimates (aggregated or as size comps, depending on "type") by year
  among several models. It uses \code{rTCSAM2013::getMDFR.Pop.Quantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, and to extract model results, and \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY} to plot them. The level of aggregation is based on the value for "type" (unlike 
\code{compreResults.Pop.Biomass}, where a cast'ing formula is specified.)
}
