% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.CohortProgression.R
\name{compareResults.Pop.CohortProgression}
\alias{compareResults.Pop.CohortProgression}
\title{Function to compare cohort progression by year among several models}
\usage{
compareResults.Pop.CohortProgression(
  objs,
  cast = "x+m+s+z",
  facet_grid = NULL,
  facet_wrap = NULL,
  dodge = 0.2,
  years = NULL,
  mxy = 4,
  nrow = 2,
  lnscale = FALSE,
  scales = "fixed",
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- casting formula for excluding x,m,s,z factor levels from sums across the unspecified factors}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{dodge}{- width to dodge overlapping series}

\item{years}{- numeric vector of years to include (or NULL for all)}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{lnscale}{- use log scale on y axis}

\item{scales}{- scales parameter for facet_grid/facet_wrap}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- filename for pdf output (or NULL for no pdf)}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares cohort progression by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced. Results are extracted using \code{rTCSAM02::getMDFR.Pop.CohortProgression},
for TCSAM02 model results (only). Plots are made using \code{plotPop.CohortProgression}.
}
