% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.GrowthMatrices.R
\name{compareResults.Pop.GrowthMatrices}
\alias{compareResults.Pop.GrowthMatrices}
\title{Compare growth matrices among several model runs}
\usage{
compareResults.Pop.GrowthMatrices(objs, showPlot = FALSE, pdf = NULL,
  verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list of ggplot objects, by sex
}
\description{
Function to compare growth matrices among several model runs.
}
\details{
If multiple models are compared, then a set of sex-specific faceted line plots 
are created. If a single model is plotted, then sex-specific bubble plots are created.
}

