% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.IF.NatZ.R
\name{compareResults.Pop.IF.NatZ}
\alias{compareResults.Pop.IF.NatZ}
\title{Function to compare initial/final population abundance estimates among several models}
\usage{
compareResults.Pop.IF.NatZ(objs, type = c("iN_yxmsz", "fN_yxmsz"),
  dodge = 0.2, showPlot = FALSE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{type}{- type of abuundance ("iN_xmsz","fN_xmsz")}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares initial/final population abundance estimates
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.PopQuantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY}.
}

