% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.MatureBiomass.R
\name{compareResults.Pop.MatureBiomass}
\alias{compareResults.Pop.MatureBiomass}
\title{Function to compare mature biomass estimates by year among several models}
\usage{
compareResults.Pop.MatureBiomass(objs, numRecent = 15, dodge = 0.2,
  showPlot = FALSE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{numRecent}{- number of "recent" years to plot}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares mature biomass estimates by year
  among several models.
}
\details{
uses \code{rTCSAM2013::getMDFR.Pop.Quantities}, 
\code{rsimTCSAM::getMDFR.Pop.Quantities}, \code{rsimTCSAM::getMDFR.Pop.Quantities}, and 
\code{plotMDFR.XY}.
}
