% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.MaturityOgives.R
\name{compareResults.Pop.MaturityOgives}
\alias{compareResults.Pop.MaturityOgives}
\title{Compare predicted maturity ogives among several model runs}
\usage{
compareResults.Pop.MaturityOgives(
  objs,
  nyrs = 10,
  plotPoints = FALSE,
  colour_scale = NULL,
  fill_scale = NULL,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{nyrs}{\itemize{
\item number of years per plot
}}

\item{plotPoints}{\itemize{
\item flag to include points (default: FALSE)
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{pdf}{\itemize{
\item name for output pdf file
}}

\item{showPlot}{\itemize{
\item flag to print plot to current device
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot object
}
\description{
Function to compare predicted maturity ogives among several model runs.
}
\details{
None.
}
