% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.NaturalMortality.BarPlot.R
\name{compareResults.Pop.NaturalMortality.BarPlot}
\alias{compareResults.Pop.NaturalMortality.BarPlot}
\title{Function to compare natural mortality rates as bars}
\usage{
compareResults.Pop.NaturalMortality.BarPlot(
  objs,
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects from TCSAM02}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object as list element
}
\description{
This function compares natural mortality estimates by
  sex and maturity state using barplots.
}
\details{
None.
}
