% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.PrM2M.R
\name{compareResults.Pop.PrM2M}
\alias{compareResults.Pop.PrM2M}
\title{Compare probability of molt-to-maturity among several model runs}
\usage{
compareResults.Pop.PrM2M(objs, dodge = 0.2, showPlot = TRUE, pdf = NULL,
  verbose = TRUE)
}
\arguments{
\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- list of resLst objects}
}
\value{
ggplot object
}
\description{
Function to compare probability of molt-to-maturity among several model runs.
}
\details{
None.
}

