% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.Recruitment.R
\name{compareResults.Pop.Recruitment}
\alias{compareResults.Pop.Recruitment}
\title{Function to compare recruitment estimates by year among several models}
\usage{
compareResults.Pop.Recruitment(objs, numRecent = 15, dodge = 0.2,
  showPlot = TRUE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{numRecent}{- number of "recent" years to plot}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares recruitment estimates by year
  among several models.
}
\details{
None.
}
