% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Pop.RecSizeDistribution.R
\name{compareResults.Pop.RecSizeDistribution}
\alias{compareResults.Pop.RecSizeDistribution}
\title{Compare recruitment size distributions among several model runs}
\usage{
compareResults.Pop.RecSizeDistribution(objs, dodge = 0.2, showPlot = TRUE,
  pdf = NULL, verbose = FALSE)
}
\arguments{
\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag to print plot to current device}

\item{pdf}{- name for output pdf file}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- list of resLst objects}
}
\value{
ggplot object
}
\description{
Function to compare recruitment size distributions among several model runs.
}
\details{
None.
}
