% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.sdRep.DerivedVars.R
\name{compareResults.sdRep.DerivedVars}
\alias{compareResults.sdRep.DerivedVars}
\title{Plot sd_report time series for recruitment and SSB from several model}
\usage{
compareResults.sdRep.DerivedVars(objs, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
list with two ggplot2 objects
}
\description{
This function plots sd_report recruitment and SSB time series from several model runs.
}
\details{
Results are extracted using \code{\link[rTCSAM02:getMDFR.SdRep.DerivedVars]{rTCSAM02::getMDFR.SdRep.DerivedVars()}} for tcsam02 model runs.
The plot objects can be re-faceted on \code{variable} as desired
(e.g., \code{p[[1]]+facet_wrap(~variable,ncol=2,scales="free")}).
}
