% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.sdRep.RecAndSSB.R
\name{compareResults.sdRep.RecAndSSB}
\alias{compareResults.sdRep.RecAndSSB}
\title{Plot sd_report time series for recruitment and SSB from several model}
\usage{
compareResults.sdRep.RecAndSSB(objs, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function plots sd_report recruitment and SSB time series from several model runs.
}
\details{
Results are extracted using \code{\link[rTCSAM02:getMDFR.SdRep.RecAndSSB]{rTCSAM02::getMDFR.SdRep.RecAndSSB()}} for tcsam02 model runs.
}
