% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.sdRep.RecAndSSB.R
\name{compareResults.sdRep.RecAndSSB}
\alias{compareResults.sdRep.RecAndSSB}
\title{Plot sd_report time series for recruitment and SSB from several model}
\usage{
compareResults.sdRep.RecAndSSB(
  objs,
  ci = 0.95,
  colour_scale = NULL,
  fill_scale = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function plots sd_report recruitment and SSB time series from several model runs.
}
\details{
Results are extracted using \code{\link[rTCSAM02:getMDFR.SdRep.RecAndSSB]{rTCSAM02::getMDFR.SdRep.RecAndSSB()}} for tcsam02 model runs.
}
