% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Abundance.R
\name{compareResults.Surveys.Abundance}
\alias{compareResults.Surveys.Abundance}
\title{Function to compare estimated survey abundance by year among several models}
\usage{
compareResults.Surveys.Abundance(objs, category = "index", cast = "x",
  years = "all", facet_grid = NULL, facet_wrap = NULL, scales = "fixed",
  dodge = 0.2, mxy = 15, nrow = 5, lnscale = FALSE, showPlot = FALSE,
  pdf = NULL, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- casting formula for excluding x,m,s,z factor levels from sums across the unspecified factors}

\item{years}{- 'all' or vector of years to include}

\item{facet_grid}{- formula for faceting using facet_grid}

\item{facet_wrap}{- formula for faceting using facet_wrap}

\item{scales}{- parameter passed to ggplot2::facet_grid()}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{nrow}{- number of rows per page, when facet_wrap'ing}

\item{lnscale}{- use log scale on y axis}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares estimated survey abundance by year
  among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}

