% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Availability.R
\name{compareResults.Surveys.Availability}
\alias{compareResults.Surveys.Availability}
\title{Function to plot survey availabilities by year using ggplot2}
\usage{
compareResults.Surveys.Availability(objs, years = "all", cast = "x",
  dodge = 0.2, facet_grid = "x~.", scales = "free_y", pdf = NULL,
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{dodge}{- width to dodge overlapping series}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{scales}{- parameter passed to ggplot2::facet_grid()}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
list of ggplot2 objects
}
\description{
This function plots survey availability estimates by year,
  sex and maturity state.
}
\details{
None.
}
