% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.AvlFcns.R
\name{compareResults.Surveys.AvlFcns}
\alias{compareResults.Surveys.AvlFcns}
\title{Function to compare survey availability functions by year among several models}
\usage{
compareResults.Surveys.AvlFcns(
  objs,
  cast = "y+x",
  fleets = "all",
  years = "all",
  dodge = 0.2,
  singlePlot = FALSE,
  mxy = 15,
  facet_wrap = NULL,
  facet_grid = ifelse(singlePlot, "x~case", "y~x"),
  pdf = NULL,
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects or dataframe from call to \code{extractMDFR.Surveys.AvlFcns}}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{fleets}{- vector of fleets to plot, or "all"}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{dodge}{- width to dodge overlapping series}

\item{singlePlot}{- flag to plot all years on single plot (be sure to adjust facet_grid)}

\item{mxy}{- max number of years per page}

\item{facet_wrap}{- ggplot2 formula to produce figure with wrapped facets}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object, or null
}
\description{
This function compares survey availability functions by year
  among several models.
}
\details{
None.
}
