% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.Biomass.R
\name{compareResults.Surveys.Biomass}
\alias{compareResults.Surveys.Biomass}
\title{Function to compare estimated survey biomass by year among several models}
\usage{
compareResults.Surveys.Biomass(
  objs,
  fleets = "all",
  cast = "y+x",
  years = "all",
  facet_grid = "x~.",
  facet_wrap = NULL,
  scales = "fixed",
  dodge = 0.2,
  mxy = 15,
  nrow = 5,
  plotPoints = FALSE,
  colour_scale = NULL,
  fill_scale = NULL,
  showPlot = FALSE,
  pdf = NULL,
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{fleets}{\itemize{
\item vector of fleets to plot, or "all"
}}

\item{cast}{\itemize{
\item cast'ing formula for aggregating by factors (y,x,m,s,z)
}}

\item{years}{\itemize{
\item 'all' or vector of years to include
}}

\item{facet_grid}{\itemize{
\item formula for faceting using facet_grid
}}

\item{facet_wrap}{\itemize{
\item formula for faceting using facet_wrap
}}

\item{scales}{\itemize{
\item parameter passed to ggplot2::facet_grid()
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{nrow}{\itemize{
\item number of rows per page, when facet_wrap'ing
}}

\item{plotPoints}{\itemize{
\item flag to include points (default: FALSE)
}}

\item{colour_scale}{\itemize{
\item ggplot2 colour scale to substitute for default (if not NULL)
}}

\item{fill_scale}{\itemize{
\item ggplot2 fill scale to substitute for default (if not NULL)
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{pdf}{\itemize{
\item creates pdf, if not NULL
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function compares estimated survey biomass by year
among several models.
}
\details{
If 'z' is a cast'ing factor, then a set of annual size composition plots are produced. Otherwise,
a set of time series plots are produced.
}
