% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.SelFcns.R
\name{compareResults.Surveys.SelFcns}
\alias{compareResults.Surveys.SelFcns}
\title{Function to compare survey selectivity functions by year among several models}
\usage{
compareResults.Surveys.SelFcns(objs, cast = "x", dodge = 0.2, mxy = 15,
  showPlot = TRUE, pdf = NULL, verbose = TRUE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{showPlot}{- flag (T/F) to show plot}

\item{pdf}{- creates pdf, if not NULL}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares survey selectivity functions by year
  among several models.
}
\details{
None.
}

