% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResults.Surveys.SelFcns.R
\name{compareResults.Surveys.SelFcns}
\alias{compareResults.Surveys.SelFcns}
\title{Function to compare survey selectivity functions by year among several models}
\usage{
compareResults.Surveys.SelFcns(objs, cast = "y+x", years = "all",
  dodge = 0.2, mxy = 15, facet_wrap = NULL, facet_grid = "y~x",
  pdf = NULL, showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{dodge}{- width to dodge overlapping series}

\item{mxy}{- max number of years per page}

\item{facet_wrap}{- ggplot2 formula to produce figure with wrapped facets}

\item{facet_grid}{- ggplot2 formula to produce figure with gridded facets}

\item{pdf}{- creates pdf, if not NULL}

\item{showPlot}{- flag (T/F) to show plot}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function compares survey selectivity functions by year
  among several models.
}
\details{
None.
}

