% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDiffs.R
\name{computeDiffs}
\alias{computeDiffs}
\title{Compute diiferences between model cases}
\usage{
computeDiffs(
  x,
  cast = "type+x+m+s+z+y",
  base = 1,
  fun = sum,
  type = c("percent", "absolute")
)
}
\arguments{
\item{x}{- dataframe to cast for differences}

\item{cast}{- cast'ing factors (left side of cast'ing equation)}

\item{base}{- name (or number) of case to use as base}

\item{fun}{- aggregation function}
}
\value{
dataframe
}
\description{
Function to compute differences between model cases
}
\details{
None.
}
