% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFits.MeanSizeComps.R
\name{extractFits.MeanSizeComps}
\alias{extractFits.MeanSizeComps}
\title{Extract fits to mean size comps by fleet among several model runs}
\usage{
extractFits.MeanSizeComps(
  objs = NULL,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  years = "all",
  ci = 0.8,
  plot1stObs = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{years}{- years to plot, as numerical vector (or "all" to plot all years)}

\item{ci}{- confidence interval}

\item{plot1stObs}{- flag (T/F) to plot observations only from first case}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}

\item{facet_grid}{- faceting formula for ggplot2::facet_grid()}

\item{pdf}{- name for output pdf file}

\item{showPlot}{- flag (T/F) to show plot}
}
\value{
dataframe in canonical format
}
\description{
Function to extract fits to mean size comps by fleet among
several model runs.
}
\details{
Uses \code{rTCSAM2013::getMDFR.SurveyQuantities()},
\code{rTCSAM2013::getMDFR.FisheryQuantities()}, \code{rTCSAM02::getMDFR.Fits.FleetData()}.
Also uses \code{reshape2::dcast}, \code{wtsUtilities::calcCIs}.
}
