% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFits.SizeComps.R
\name{extractFits.SizeComps}
\alias{extractFits.SizeComps}
\title{Extract fits to size comps by fleet from several model runs}
\usage{
extractFits.SizeComps(
  objs = NULL,
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  years = "all",
  plot1stObs = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{catch.type}{- catch type ('index','retained',  or 'total')}

\item{years}{- years to plot, as numerical vector (or "all" to plot all years)}

\item{plot1stObs}{- flag (T/F) to plot observations only from first case}

\item{verbose}{- flag (T/F) to print diagnostic information}

\item{obj}{- object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects}
}
\value{
dataframe in canonical format
}
\description{
Function to extract fits to size comps by fleet from
several model runs.
}
\details{
Uses \code{rTCSAM2013::getMDFR.SurveyQuantities()},
\code{rTCSAM2013::getMDFR.FisheryQuantities()}, \code{rTCSAM02::getMDFR.Fits.FleetData()}.
}
