% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.CatchAbundance.R
\name{extractMDFR.Fisheries.CatchAbundance}
\alias{extractMDFR.Fisheries.CatchAbundance}
\title{Function to extract estimated fishery catch abundance by year among several models as a dataframe}
\usage{
extractMDFR.Fisheries.CatchAbundance(
  objs,
  category = c("captured", "discarded", "retained", "discard mortality", "index"),
  cast = NULL,
  years = "all",
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{category}{- 'captured','discarded','retained','discard mortality', or 'index'}

\item{cast}{- cast'ing formula for aggregating by factors (y,x,m,s,z)}

\item{years}{- 'all' or vector of years to include}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
This function extracts estimated fishery catch abundance by year
  among several models as a dataframe.
}
\details{
None.
}
