% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.CatchBiomass.R
\name{extractMDFR.Fisheries.CatchBiomass}
\alias{extractMDFR.Fisheries.CatchBiomass}
\title{Function to extract estimated fishery catch biomass by year among several models as a dataframe}
\usage{
extractMDFR.Fisheries.CatchBiomass(objs, category = c("captured", "discarded",
  "retained", "discard mortality", "index"), cast = NULL, years = "all",
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{category}{- 'captured','discarded','retained','discard mortality', or 'index'}

\item{cast}{- cast'ing formula for aggregating by factors (x,m,s,z)}

\item{years}{- 'all' or vector of years to include}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts estimated fishery catch biomass by year
  among several models as a dataframe.
}
\details{
None.
}
