% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fisheries.RetFcns.R
\name{extractMDFR.Fisheries.RetFcns}
\alias{extractMDFR.Fisheries.RetFcns}
\title{Function to extract fishery retention functions by year among several models}
\usage{
extractMDFR.Fisheries.RetFcns(objs, cast = "y+x", years = "all",
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
This function extracts fishery retention functions by year
  among several models.
}
\details{
None.
}

