% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.AbundanceData.R
\name{extractMDFR.Fits.AbundanceData}
\alias{extractMDFR.Fits.AbundanceData}
\title{Extract fits to abundance time series by fleet among several model runs}
\usage{
extractMDFR.Fits.AbundanceData(
  objs = NULL,
  fleets = "all",
  fleet.type = c("survey", "fishery"),
  catch.type = c("index", "retained", "discard", "total"),
  ci = 0.8,
  fishery.pdfType = c("norm2", "normal", "lognormal"),
  verbose = FALSE
)
}
\arguments{
\item{fleets}{\itemize{
\item names of fleets to include (or "all")
}}

\item{fleet.type}{\itemize{
\item fleet type ('fishery' or 'survey')
}}

\item{catch.type}{\itemize{
\item catch type ('index','retained',  or 'total')
}}

\item{ci}{\itemize{
\item confidence interval for plots
}}

\item{fishery.pdfType}{\itemize{
\item assumed distribution for fishery data
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}

\item{obj}{\itemize{
\item object that can be converted into a list of tcsam2013.resLst and/or tcsam02.resLst objects
}}
}
\value{
dataframe
}
\description{
Function to extract fits to abundance time series by fleet among
several model runs.
}
\details{
Uses \code{\link[rTCSAM2013:getMDFR.SurveyQuantities]{rTCSAM2013::getMDFR.SurveyQuantities()}},
\code{\link[rTCSAM2013:getMDFR.FisheryQuantities]{rTCSAM2013::getMDFR.FisheryQuantities()}}, \code{\link[rTCSAM02:getMDFR.Fits.FleetData]{rTCSAM02::getMDFR.Fits.FleetData()}}.
}
