% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Fits.EffectiveNs.R
\name{extractMDFR.Fits.EffectiveNs}
\alias{extractMDFR.Fits.EffectiveNs}
\title{Function to extract effective Ns from size comps by year among several models}
\usage{
extractMDFR.Fits.EffectiveNs(objs, fleet.type = c("survey", "fishery"),
  category = c("index", "captured", "discarded", "retained",
  "discard mortality"), verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{fleet.type}{- 'survey','fishery'}

\item{category}{- 'total','discard','retained','discard mortality', or 'index'}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts effective Ns from size comps by year
  among several models.
}
\details{
Results are extracted using \code{rTCSAM02::getMDFR.Fits.EffectiveNs}, as appropriate.
}
