% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.Biomass.R
\name{extractMDFR.Surveys.Biomass}
\alias{extractMDFR.Surveys.Biomass}
\title{Function to compare estimated survey biomass by year among several models}
\usage{
extractMDFR.Surveys.Biomass(objs, category = "index", cast = "y+x",
  years = "all", verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- cast'ing formula for aggregating by factors (y,x,m,s,z)}

\item{years}{- 'all' or vector of years to include}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format.
}
\description{
This function extracts estimated survey biomass by year
  among several models as a dataframe
}
\details{
None.
}
