% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.Catchability.R
\name{extractMDFR.Surveys.Catchability}
\alias{extractMDFR.Surveys.Catchability}
\title{Function to extract survey catchabilities}
\usage{
extractMDFR.Surveys.Catchability(objs, cast = "x", verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over (x,m,s)}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe in canonical format
}
\description{
This function extracts survey catchability estimates by year,
  sex and maturity state.
}
\details{
None.
}

