% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMDFR.Surveys.SelFcns.R
\name{extractMDFR.Surveys.SelFcns}
\alias{extractMDFR.Surveys.SelFcns}
\title{Function to extract survey selectivity functions by year among several models}
\usage{
extractMDFR.Surveys.SelFcns(objs, cast = "y+x", years = "all",
  verbose = FALSE)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{cast}{- formula to exclude factors from "averaging" over}

\item{years}{- vector of years to show, or 'all' to show all years}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
dataframe
}
\description{
This function extracts survey selectivity functions by year
  among several models.
}
\details{
None.
}

