% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.Characteristics.Fisheries.R
\name{modelComparisons.Characteristics.Fisheries}
\alias{modelComparisons.Characteristics.Fisheries}
\title{Render a document of comparison plots of estimated fishery characteristics from several models}
\usage{
modelComparisons.Characteristics.Fisheries(models, fisheries = "all",
  selyears = "all", retyears = "all",
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.EstimatedCharacteristics.Fisheries.Rmd",
  package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package =
  "wtsUtilities"), pdf_styles = system.file("rmd/StylesForRmdPDFs.sty",
  package = "wtsUtilities"), clean = FALSE)
}
\arguments{
\item{models}{- named list of model results (as resLst objects) to compare}

\item{fisheries}{- vector of fisheries to plot, or "all"}

\item{selyears}{- vector of years to show selectivity functions for, or 'all' to show all years}

\item{retyears}{- vector of years to show retention functions for, or 'all' to show all years}

\item{output_format}{- "word_document" or "pdf_document"}

\item{output_dir}{- path to folder to use for output}

\item{rmd}{- Rmd file to process (defalut="modelComparisons.Rmd")}

\item{docx_styles}{- full path to Word (docx) style template for Word documents}

\item{pdf_styles}{- full path to style template for pdf documents}

\item{clean}{- T/F to delete intermediate files}

\item{rmd_dir}{- folder enclosing rmd file}
}
\description{
Function to render a document of comparison plots of estimated fishery
characteristics from several models.
}
\details{
The default resulting document title will be of the form "ModelComparisons.EstimatedCharacteristics.Fisheries.mmm.ext",
where "ext" is the appropriate file extension and "mmm" is a dash-separated string of model names.
}
