% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparisons.InputData.R,
%   R/modelComparisons.ManagementQuantities.R, R/modelComparisons.R
\name{modelComparisons}
\alias{modelComparisons}
\title{Render a document of model comparison plots}
\usage{
modelComparisons(
  models,
  includeManagementQuantities = FALSE,
  ManagementQuantities = list(base = 1, verbose = FALSE),
  includeInputData = FALSE,
  InputData = list(includeSurveyData = TRUE, surveys = list(plotAbundance = TRUE,
    plotBiomass = TRUE, plotZCs = TRUE, fleets = "all", ci = 0.8, numRecent = 30),
    includeFisheryData = TRUE, fisheries = list(fleets = "all", retained =
    list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8, numRecent =
    30), total = list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8,
    numRecent = 30), effort = list(plot = TRUE, numRecent = 30)), includeGrowthData =
    TRUE, includeMaturityOgiveData = TRUE, 
     verbose = FALSE),
  includeParameterTables = FALSE,
  ParameterTables = list(categories = "all", processes = "all", types = "all", params =
    "all", verbose = FALSE),
  includePopProcesses = FALSE,
  PopProcesses = list(verbose = FALSE),
  includePopQuantities = FALSE,
  PopQuantities = list(verbose = FALSE),
  includeModelFitsToOtherData = FALSE,
  ModelFitsToOtherData = list(plot1stObs = TRUE, minSizeForMaturityData = 60, verbose =
    FALSE),
  includeModelFitsACD = FALSE,
  ModelFitsACD = list(plot1stObs = TRUE, numRecent = 30, verbose = FALSE),
  includeModelFitsZCs = FALSE,
  ModelFitsZCs = list(plot1stObs = TRUE, fleets = "all", verbose = FALSE),
  includeCharacteristicsForSurveys = FALSE,
  CharacteristicsForSurveys = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    avlyears = "all", capyears = "all", mxy = 5, verbose = FALSE),
  includeCharacteristicsForFisheries = FALSE,
  CharacteristicsForFisheries = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    retyears = "all", verbose = FALSE),
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.Rmd", package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)

modelComparisons(
  models,
  includeManagementQuantities = FALSE,
  ManagementQuantities = list(base = 1, verbose = FALSE),
  includeInputData = FALSE,
  InputData = list(includeSurveyData = TRUE, surveys = list(plotAbundance = TRUE,
    plotBiomass = TRUE, plotZCs = TRUE, fleets = "all", ci = 0.8, numRecent = 30),
    includeFisheryData = TRUE, fisheries = list(fleets = "all", retained =
    list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8, numRecent =
    30), total = list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8,
    numRecent = 30), effort = list(plot = TRUE, numRecent = 30)), includeGrowthData =
    TRUE, includeMaturityOgiveData = TRUE, 
     verbose = FALSE),
  includeParameterTables = FALSE,
  ParameterTables = list(categories = "all", processes = "all", types = "all", params =
    "all", verbose = FALSE),
  includePopProcesses = FALSE,
  PopProcesses = list(verbose = FALSE),
  includePopQuantities = FALSE,
  PopQuantities = list(verbose = FALSE),
  includeModelFitsToOtherData = FALSE,
  ModelFitsToOtherData = list(plot1stObs = TRUE, minSizeForMaturityData = 60, verbose =
    FALSE),
  includeModelFitsACD = FALSE,
  ModelFitsACD = list(plot1stObs = TRUE, numRecent = 30, verbose = FALSE),
  includeModelFitsZCs = FALSE,
  ModelFitsZCs = list(plot1stObs = TRUE, fleets = "all", verbose = FALSE),
  includeCharacteristicsForSurveys = FALSE,
  CharacteristicsForSurveys = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    avlyears = "all", capyears = "all", mxy = 5, verbose = FALSE),
  includeCharacteristicsForFisheries = FALSE,
  CharacteristicsForFisheries = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    retyears = "all", verbose = FALSE),
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.Rmd", package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)

modelComparisons(
  models,
  includeManagementQuantities = FALSE,
  ManagementQuantities = list(base = 1, verbose = FALSE),
  includeInputData = FALSE,
  InputData = list(includeSurveyData = TRUE, surveys = list(plotAbundance = TRUE,
    plotBiomass = TRUE, plotZCs = TRUE, fleets = "all", ci = 0.8, numRecent = 30),
    includeFisheryData = TRUE, fisheries = list(fleets = "all", retained =
    list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8, numRecent =
    30), total = list(plotAbundance = TRUE, plotBiomass = TRUE, plotZCs = TRUE, ci = 0.8,
    numRecent = 30), effort = list(plot = TRUE, numRecent = 30)), includeGrowthData =
    TRUE, includeMaturityOgiveData = TRUE, 
     verbose = FALSE),
  includeParameterTables = FALSE,
  ParameterTables = list(categories = "all", processes = "all", types = "all", params =
    "all", verbose = FALSE),
  includePopProcesses = FALSE,
  PopProcesses = list(verbose = FALSE),
  includePopQuantities = FALSE,
  PopQuantities = list(verbose = FALSE),
  includeModelFitsToOtherData = FALSE,
  ModelFitsToOtherData = list(plot1stObs = TRUE, minSizeForMaturityData = 60, verbose =
    FALSE),
  includeModelFitsACD = FALSE,
  ModelFitsACD = list(plot1stObs = TRUE, numRecent = 30, verbose = FALSE),
  includeModelFitsZCs = FALSE,
  ModelFitsZCs = list(plot1stObs = TRUE, fleets = "all", verbose = FALSE),
  includeCharacteristicsForSurveys = FALSE,
  CharacteristicsForSurveys = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    avlyears = "all", capyears = "all", mxy = 5, verbose = FALSE),
  includeCharacteristicsForFisheries = FALSE,
  CharacteristicsForFisheries = list(plot1stObs = TRUE, fleets = "all", selyears = "all",
    retyears = "all", verbose = FALSE),
  output_format = c("word_document", "pdf_document"),
  output_dir = getwd(),
  rmd = system.file("rmd/modelComparisons.Rmd", package = "rCompTCMs"),
  docx_styles = system.file("rmd/StylesForRmdDocs.docx", package = "wtsUtilities"),
  pdf_styles = system.file("rmd/StylesForRmdPDFs.sty", package = "wtsUtilities"),
  clean = FALSE
)
}
\arguments{
\item{models}{\itemize{
\item named list of model results (as resLst objects) to compare
}}

\item{includeManagementQuantities}{\itemize{
\item flag to include comparison of management quantities
}}

\item{ManagementQuantities}{\itemize{
\item list controlling comparison of management quantities (see \code{\link[=modelComparisons.ManagementQuantities]{modelComparisons.ManagementQuantities()}})
}}

\item{includeInputData}{\itemize{
\item flag (T/F) to include section
}}

\item{InputData}{\itemize{
\item list controlling what gets plotted
}}

\item{includeParameterTables}{\itemize{
\item flag (T/F) to include section
}}

\item{ParameterTables}{\itemize{
\item list(verbose=FALSE)  (see \code{\link[=modelComparisons.ParameterTables]{modelComparisons.ParameterTables()}})
}}

\item{includePopProcesses}{\itemize{
\item flag (T/F) to include section
}}

\item{PopProcesses}{\itemize{
\item list(verbose=FALSE) (see \code{\link[=modelComparisons.PopProcesses]{modelComparisons.PopProcesses()}})
}}

\item{includePopQuantities}{\itemize{
\item flag (T/F) to include section
}}

\item{PopQuantities}{\itemize{
\item list(verbose=FALSE)  (see \code{\link[=modelComparisons.PopQuantitiess]{modelComparisons.PopQuantitiess()}})
}}

\item{includeModelFitsToOtherData}{\itemize{
\item flag (T/F) to include section
}}

\item{ModelFitsToOtherData}{\itemize{
\item list(plot1stObs=TRUE,minSizeForMaturityData=60,verbose=FALSE)  (see \code{\link[=modelComparisons.ModelFitsToOtherData]{modelComparisons.ModelFitsToOtherData()}})
}}

\item{includeModelFitsACD}{\itemize{
\item flag (T/F) to include section
}}

\item{ModelFitsACD}{\itemize{
\item list(plot1stObs=TRUE,numRecent=30,verbose=FALSE) (see \code{\link[=modelComparisons.ModelFitsACD]{modelComparisons.ModelFitsACD()}})
}}

\item{includeModelFitsZCs}{\itemize{
\item flag (T/F) to include section
}}

\item{ModelFitsZCs}{\itemize{
\item list(plot1stObs=TRUE,fleets="all",verbose=FALSE) (see \code{\link[=modelComparisons.ModelFitsZCs]{modelComparisons.ModelFitsZCs()}})
}}

\item{includeCharacteristicsForSurveys}{\itemize{
\item flag (T/F) to include section
}}

\item{CharacteristicsForSurveys}{\itemize{
\item list(plot1stObs=TRUE,fleets="all",selyears="all",avlyears="all",capyears="all",mxy=5,verbose=FALSE)  (see \code{\link[=modelComparisons.CharacteristicsSurveys]{modelComparisons.CharacteristicsSurveys()}})
}}

\item{includeCharacteristicsForFisheries}{\itemize{
\item flag (T/F) to include section
}}

\item{CharacteristicsForFisheries}{\itemize{
\item list(plot1stObs=TRUE,fleets="all",selyears="all",retyears="all",verbose=FALSE) (see \code{\link[=modelComparisons.CharacteristicsFisheries]{modelComparisons.CharacteristicsFisheries()}})
}}

\item{output_format}{\itemize{
\item "word_document" or "pdf_document"
}}

\item{output_dir}{\itemize{
\item path to folder to use for output
}}

\item{rmd}{\itemize{
\item Rmd file to process (defalut=system.file("rmd/modelComparisons.Rmd",package="rCompTCMs"))
}}

\item{docx_styles}{\itemize{
\item full path to Word (docx) style template for Word documents
}}

\item{pdf_styles}{\itemize{
\item full path to style template for pdf documents
}}

\item{clean}{\itemize{
\item T/F to delete intermediate files
}}

\item{plot1stObs}{\itemize{
\item flag to plot observed data only from first model
}}

\item{rmd_dir}{\itemize{
\item folder enclosing rmd file
}}

\item{base}{\itemize{
\item name or index of base case for difference calculations
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic output
}}
}
\description{
Function to render a document of model comparison plots

Function to render a document of model comparison plots for various management quantities

Function to render a document of model comparison plots
}
\details{
Resulting document title will be of the form "ModelComparisons.mmm.ext", where "ext" is the appropriate
file extension and "mmm" is a dash-separated string of model names.

Resulting document title will be of the form "ModelComparisons.ManagementQuantities.mmm.ext", where "ext" is the appropriate
file extension and "mmm" is a dash-separated string of model names.

Resulting document title will be of the form "ModelComparisons.mmm.ext", where "ext" is the appropriate
file extension and "mmm" is a dash-separated string of model names.
}
