% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.ACD}
\alias{plotLollipops.ACD}
\title{plot lollipops for fits to aggregated catch data}
\usage{
plotLollipops.ACD(
  mdfr,
  type = "z-score",
  extreme = 3,
  dw = 1,
  size = 0.1,
  plotPoints = TRUE,
  xlab = "year",
  ylab = type
)
}
\arguments{
\item{mdfr}{\itemize{
\item input data
}}

\item{type}{\itemize{
\item type of value to plot (e.g., "z-score")
}}

\item{extreme}{\itemize{
\item criteria for extreme values
}}

\item{dw}{\itemize{
\item dodge width (x-axis units)
}}

\item{size}{\itemize{
\item line size for lollipop stems
}}

\item{plotPoints}{\itemize{
\item flag to plot "pop" part of lolli
}}

\item{xlab}{\itemize{
\item x-axis label
}}

\item{ylab}{\itemize{
\item y-axis label
}}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to aggregated catch data.
}
\details{
Under construction!
}
