% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.ACD}
\alias{plotLollipops.ACD}
\title{plot lollipops for fits to size composition data}
\usage{
plotLollipops.ACD(
  mdfr,
  type = "z-score",
  extreme = 3,
  dw = 1,
  xlab = "year",
  ylab = type
)
}
\arguments{
\item{mdfr}{- input data}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to size composition data.
}
\details{
Under construction!
}
