% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.GrowthData}
\alias{plotLollipops.GrowthData}
\title{plot lollipops for fits to molt increment data}
\usage{
plotLollipops.GrowthData(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 1,
  size = 0.1,
  xlab = "pre-molt size",
  ylab = type
)
}
\arguments{
\item{mdfr}{- input dataframe from [rCompTCMs::extractMDFR.Fits.GrowthData()]}

\item{type}{- type of value to plot (e.g., "zscores")}

\item{extreme}{- criteria for extreme values}

\item{dw}{- dodge width (x-axis units)}

\item{size}{- line size for lollipop stems}

\item{xlab}{- x-axis label}

\item{ylab}{- y-axis label}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to molt increment data.
}
\details{
Plots \code{type} as lollipops against pre-molt size.
}
