% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.GrowthData}
\alias{plotLollipops.GrowthData}
\title{plot lollipops for fits to molt increment data}
\usage{
plotLollipops.GrowthData(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 1,
  size = 0.1,
  xlab = "pre-molt size",
  ylab = type
)
}
\arguments{
\item{mdfr}{\itemize{
\item input dataframe from \code{\link[=extractMDFR.Fits.GrowthData]{extractMDFR.Fits.GrowthData()}}
}}

\item{type}{\itemize{
\item type of value to plot (e.g., "zscores")
}}

\item{extreme}{\itemize{
\item criteria for extreme values
}}

\item{dw}{\itemize{
\item dodge width (x-axis units)
}}

\item{size}{\itemize{
\item line size for lollipop stems
}}

\item{xlab}{\itemize{
\item x-axis label
}}

\item{ylab}{\itemize{
\item y-axis label
}}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to molt increment data.
}
\details{
Plots \code{type} as lollipops against pre-molt size.
}
