% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.GrowthData}
\alias{plotLollipops.GrowthData}
\title{plot lollipops for fits to molt increment data}
\usage{
plotLollipops.GrowthData(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 0,
  xlab = "pre-molt size",
  ylab = type
)
}
\arguments{
\item{mdfr}{- input dataframe from [rCompTCMs::extractMDFR.Fits.GrowthData()]}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to molt increment data.
}
\details{
Plots \code{type} as lollipops against pre-molt size.
}
