% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.MaturityOgiveData}
\alias{plotLollipops.MaturityOgiveData}
\title{plot lollipops for fits to maturity ogive data}
\usage{
plotLollipops.MaturityOgiveData(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 5,
  size = 0.1,
  plotPoints = TRUE,
  nrow = 5,
  xlab = "size (mm CW)",
  ylab = type
)
}
\arguments{
\item{mdfr}{\itemize{
\item input dataframe from \code{\link[=extractMDFR.Fits.MaturityOgiveData]{extractMDFR.Fits.MaturityOgiveData()}}
}}

\item{type}{\itemize{
\item type of value to plot (e.g., "zscores")
}}

\item{extreme}{\itemize{
\item criteria for extreme values
}}

\item{dw}{\itemize{
\item dodge width (x-axis units)
}}

\item{size}{\itemize{
\item line size for lollipop stems
}}

\item{plotPoints}{\itemize{
\item flag to plot "pop" part of lolli
}}

\item{nrow}{\itemize{
\item number of rows for faceting
}}

\item{xlab}{\itemize{
\item x-axis label
}}

\item{ylab}{\itemize{
\item y-axis label
}}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to maturity ogive data.
}
\details{
Plots \code{type} as lollipops against size, faceted by year.
}
