% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.ZCs}
\alias{plotLollipops.ZCs}
\title{plot lollipops for fits to size composition data}
\usage{
plotLollipops.ZCs(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 5,
  xlab = "size bin (mm CW)",
  ylab = "z-score",
  facet_wrap = ~y,
  ncol = 2
)
}
\arguments{
\item{mdfr}{- input data}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to size composition data.
}
\details{
Under construction!
}
