% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLollipops.R
\name{plotLollipops.ZCs}
\alias{plotLollipops.ZCs}
\title{plot lollipops for fits to size composition data}
\usage{
plotLollipops.ZCs(
  mdfr,
  type = "zscores",
  extreme = 3,
  dw = 5,
  size = 0.1,
  plotPoints = TRUE,
  xlab = "size bin (mm CW)",
  ylab = "z-score",
  facet_wrap = ~y,
  ncol = 2
)
}
\arguments{
\item{mdfr}{\itemize{
\item input data
}}

\item{type}{\itemize{
\item type of value to plot (e.g., "zscores")
}}

\item{extreme}{\itemize{
\item criteria for extreme values
}}

\item{dw}{\itemize{
\item dodge width (x-axis units)
}}

\item{size}{\itemize{
\item line size for lollipop stems
}}

\item{plotPoints}{\itemize{
\item flag to plot "pop" part of lolli
}}

\item{xlab}{\itemize{
\item x-axis label
}}

\item{ylab}{\itemize{
\item y-axis label
}}
}
\value{
ggplot2 plot object
}
\description{
function to plot lollipops for fits to size composition data.
}
\details{
Under construction!
}
