% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Fits.TimeSeries.R
\name{plotMDFR.Fits.TimeSeries}
\alias{plotMDFR.Fits.TimeSeries}
\title{Plot comparisons of fits to time series from a set of model runs, with a "zoom" for recent years}
\usage{
plotMDFR.Fits.TimeSeries(
  dfr,
  plot1stObs = TRUE,
  numRecent = 15,
  x = "y",
  y = "val",
  lci = "lci",
  uci = "uci",
  case = "case",
  type = "type",
  facets = NULL,
  scales = "fixed",
  position = ggplot2::position_dodge(0.2),
  plotObs = TRUE,
  plotMod = TRUE,
  xlab = "year",
  ylab = NULL,
  title = NULL,
  xlims = NULL,
  ylims = NULL,
  colour_scale = ggplot2::scale_color_hue(),
  showPlot = FALSE
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe
}}

\item{plot1stObs}{\itemize{
\item flag to plot observations from first case only
}}

\item{numRecent}{\itemize{
\item number of "recent" years to plot
}}

\item{x}{\itemize{
\item column name with x axis values
}}

\item{y}{\itemize{
\item column name with y axis values
}}

\item{lci}{\itemize{
\item column name with y axis values
}}

\item{uci}{\itemize{
\item column name with y axis values
}}

\item{case}{\itemize{
\item column name with case names
}}

\item{type}{\itemize{
\item column name with type values (i.e., "observed","predicted")
}}

\item{facets}{\itemize{
\item grid faceting formula (as an expresssion for multilevel faceting)
}}

\item{scales}{\itemize{
\item ggplot2 scales option for facet_grid
}}

\item{position}{\itemize{
\item indicates ggplot2 position_ to use ('dodge','jitter','identity',)
}}

\item{plotObs}{\itemize{
\item plot observations
}}

\item{plotMod}{\itemize{
\item plot model fits/predictions
}}

\item{xlab}{\itemize{
\item 
}}

\item{ylab}{\itemize{
\item 
}}

\item{title}{\itemize{
\item 
}}

\item{xlims}{\itemize{
\item 
}}

\item{ylims}{\itemize{
\item 
}}

\item{colour_scale}{\itemize{
\item ggplot2 scale_colour object (default is ggplot2::scale_colour_hue())
}}

\item{showPlot}{\itemize{
\item flag (T/F) to print plot
}}
}
\value{
list with requested ggplot objects
}
\description{
Function to plot a comparison of fits to time series from a set of model runs, with a "zoom" for recent years.
}
\details{
Uses \code{\link[=plotMDFR.Fits.TimeSeries1]{plotMDFR.Fits.TimeSeries1()}} to make the plots. \code{numRecent} provides the "zoom"
for a second set of faceted plots including only the most recent years. If \code{numRecent} is 0,
this plot is not created and an NA is substituted.

Note that \code{facets} should be given as an expression, not as a character string, if you want multilevel faceting.
}
