% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.CohortProgression.R
\name{plotPop.CohortProgression}
\alias{plotPop.CohortProgression}
\title{Function to plot cohort progression by year among several models}
\usage{
plotPop.CohortProgression(
  mdfr,
  facet_grid = NULL,
  facet_wrap = NULL,
  dodge = 0.2,
  years = NULL,
  mxy = 4,
  nrow = 2,
  lnscale = FALSE,
  scales = "fixed",
  types = c("progression", "byyear"),
  plotPoints = FALSE,
  shapes = "s",
  plotLines = TRUE,
  linetypes = "ms",
  showPlot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{\itemize{
\item melted data.frame with cohort progressions
}}

\item{facet_grid}{\itemize{
\item formula for faceting using facet_grid
}}

\item{facet_wrap}{\itemize{
\item formula for faceting using facet_wrap
}}

\item{dodge}{\itemize{
\item width to dodge overlapping series
}}

\item{years}{\itemize{
\item vector of years to include in plots (or NULL for all)
}}

\item{mxy}{\itemize{
\item max number of years per page
}}

\item{nrow}{\itemize{
\item number of rows per page, when facet_wrap'ing
}}

\item{lnscale}{\itemize{
\item use log scale on y axis
}}

\item{scales}{\itemize{
\item scales parameter for facet_grid/facet_wrap
}}

\item{types}{\itemize{
\item cohort plot types ("progression","byyear")
}}

\item{plotPoints}{\itemize{
\item flag to plot points
}}

\item{shapes}{\itemize{
\item name of column for ggplot shape factors (defaults to 's' for shell condition)
}}

\item{plotLines}{\itemize{
\item flag to plot cohorts as lines
}}

\item{linetypes}{\itemize{
\item name of column for ggplot linetype factors (defaults to 'ms' for maturity + shell condition)
}}

\item{showPlot}{\itemize{
\item flag (T/F) to show plot
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
a list of ggplot2 objects
}
\description{
This function plots cohort progression by year
among several models.
}
\details{
Can plot cohort abundance as a progression by stage for each model or
as size distributions by year with colours indicating different models.
}
