% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.MeanGrowth.R
\name{plotPop.MeanGrowth}
\alias{plotPop.MeanGrowth}
\title{Plot mean post-molt size for several model runs}
\usage{
plotPop.MeanGrowth(mdfr, dodge = 0.2, showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{dodge}{- width to dodge overlapping series}

\item{showPlot}{- flag to print plot to current device}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot object
}
\description{
Function to plot mean post-molt size for several model runs.
}
\details{
None.
}
