% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.PrM2M.R
\name{plotPop.PrM2M}
\alias{plotPop.PrM2M}
\title{Function to plot probability of molt-to-maturity for several model runs}
\usage{
plotPop.PrM2M(
  mdfr,
  dodge = 0.2,
  xbnds = NULL,
  showPlot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{dodge}{- width to dodge overlapping series}

\item{xbnds}{- two-element vector with x-axis bound (or NULL)}

\item{showPlot}{- flag to print plot to current device}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot object
}
\description{
Function to plot probability of molt-to-maturity for several model runs.
}
\details{
None.
}
