% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.PrMolt.R
\name{plotPop.PrMolt}
\alias{plotPop.PrMolt}
\title{Function to plot annual molt probability for immature crab from several model runs}
\usage{
plotPop.PrMolt(
  mdfr,
  dodge = 0.2,
  xbnds = NULL,
  showPlot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{dodge}{- width to dodge overlapping series}

\item{xbnds}{- two-element vector with x-axis bound (or NULL)}

\item{showPlot}{- flag to print plot to current device}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot object
}
\description{
Function to plot annual molt probability for immature crab from several model runs.
}
\details{
None.
}
