% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.ParametersAtBounds.R
\name{printResults.ParametersAtBounds}
\alias{printResults.ParametersAtBounds}
\title{Function to print a table of parameters at bounds for several models}
\usage{
printResults.ParametersAtBounds(
  objs = NULL,
  mdfr = NULL,
  delta = 1e-04,
  format = c("latex", "docx"),
  landscape = FALSE,
  fontsize = "11pt",
  verbose = FALSE
)
}
\arguments{
\item{objs}{\itemize{
\item list of resLst objects
}}

\item{mdfr}{\itemize{
\item dataframe from call to \code{extractMDFR.Results.ParametersAtBounds} (as alternative to objs)
}}

\item{delta}{\itemize{
\item fraction of range which defines "at the bounds"
}}

\item{format}{\itemize{
\item 'latex' or 'docx'
}}

\item{landscape}{\itemize{
\item flag indicating whether page orientation is landscape or portrait
}}

\item{fontsize}{\itemize{
\item font size (can be 'small')
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
ggplot2 object
}
\description{
This function prints a table of parameters at bounds for several models.
}
\details{
Uses \code{extractMDFR.Results.ParametersAtBounds} to get a dataframe of parameters at one of their bounds,
based upon the value for delta. If format="latex", the package \code{tables} is used to create and print a table
in latex format.
}
