% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.ParameterValues.R
\name{printResults.ParameterValues}
\alias{printResults.ParameterValues}
\title{Function to print a table of parameters from several models}
\usage{
printResults.ParameterValues(
  objs = NULL,
  mdfr = NULL,
  format = c("latex", "docx"),
  landscape = FALSE,
  fontsize = "11pt",
  verbose = FALSE
)
}
\arguments{
\item{objs}{- list of resLst objects}

\item{mdfr}{- dataframe from call to \code{extractMDFR.Results.ParameterValues} (as alternative to objs)}

\item{format}{- 'latex' or 'docx'}

\item{landscape}{- flag indicating whether page orientation is landscape or portrait}

\item{fontsize}{- font size (can be 'small')}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
ggplot2 object
}
\description{
This function prints a table of parameters from several models.
}
\details{
Uses \code{extractMDFR.Results.ParameterValues} to get a dataframe of parameters.
If format="latex", the package \code{tables} is used to create and print a table
in latex format. Output format for values and standard deviations is 3 significant figures.
}
