% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroFunctions.R
\name{retroGetMDFR}
\alias{retroGetMDFR}
\title{Get a melted dataframe of model estimates from a list of retrospective model results}
\usage{
retroGetMDFR(modsRetro, FUN, yadj = 0, ...)
}
\arguments{
\item{modsRetro}{- named list of resLst objects, with names corresponding to retrospective peels}

\item{FUN}{- rCompTCMs function used to extract the desired estimates (i.e., an extractMDFR function)}

\item{yadj}{- adjustment to make to model years}

\item{...}{- additional arguments passed to FUN}
}
\value{
dataframe
}
\description{
This function extracts a melted dataframe from a list of retrospective model results.
}
\details{
Results are extracted using \code{FUN} for a retrospective series of tcsam02 model runs.
}
