% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroPlots.R
\name{retroPlotFitsAsTimeSeries}
\alias{retroPlotFitsAsTimeSeries}
\title{Create a plot of model fits to time series from a retrospective analysis}
\usage{
retroPlotFitsAsTimeSeries(
  dfr,
  ylab,
  units,
  lnscale = FALSE,
  dodge = 0,
  size = 2,
  colour = "case",
  shape = colour,
  facet_grid = termyear ~ .,
  verbose = FALSE
)
}
\arguments{
\item{dfr}{- dataframe of estimated values extracted from a series of retrospective model runs}

\item{ylab}{- label for y-axis}

\item{units}{- label for y-axis units}

\item{lnscale}{- flag to use log scale}

\item{dodge}{- amount to dodge overlapping values}

\item{size}{- point shape size}

\item{colour}{- column name indicating retrospective model (default='case')}

\item{shape}{- column name for shapes (default=value assigned to colour)}

\item{facet_grid}{- formula to apply faceting}

\item{verbose}{- flag}
}
\value{
a ggplot2 object
}
\description{
This function creates a plot of model fits to time series from a retrospective analysis.
}
\details{
Uses \code{rCompTCMs::plotMDFR.XY} to plot the timeseries.
}
