% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroPlots.R
\name{retroPlotOFLResults}
\alias{retroPlotOFLResults}
\title{Create plots of several OFL-related quantities from a retrospective analysis}
\usage{
retroPlotOFLResults(dfr.OFLs, x = "case", xlab = "Peel")
}
\arguments{
\item{dfr.OFLs}{- dataframe of OFL-related values from a retrospective analysis}

\item{x}{- name of dataframe column with "peel"labels}

\item{xlab}{- label for x-axis}
}
\value{
list of ggplot2 objects
}
\description{
This function create plots for several OFL-related quantities from a retrospective analysis.
}
\details{
Creates bar plots for average recruitment, Bmsy, Fmsy, terminal year MMB,
projected MMB, and status
}
