% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroPlots.R
\name{retroPlotValuesBySize}
\alias{retroPlotValuesBySize}
\title{Create a plot of values across size from a retrospective analysis}
\usage{
retroPlotValuesBySize(dfr, xlab = "size (mm CW)", ylab = "", facet_grid = NULL)
}
\arguments{
\item{dfr}{- dataframe of estimated values across size extracted from a series of retrospective model runs}

\item{xlab}{- label for x-axis}

\item{ylab}{- label for y-axis}

\item{facet_grid}{- formula to apply faceting}
}
\value{
a ggplot2 object
}
\description{
This function creates a plot of values across size from a retrospective analysis.
}
\details{
The column "case" should indicate the peels of the retrospective runs
}
